refactor(main): move nine subsystems out of main.rs (7,006 → 3,840 lines)
main.rs was still ~15 unrelated subsystems sharing one file. Nine of them were leaves —
they call into the crate but almost nothing in main.rs calls back — so each moves out with
one glob re-export and no behaviour change:

  src/cli/agents_cmd.rs    663   `aizen agents …` + the install machinery
  src/cli/time.rs          590   `aizen time …` / `/timemachine`
  src/cli/apps.rs          493   `aizen apps …` (MCP registry)
  src/cli/sessions.rs      373   `/sessions`, `/import`, `aizen import`
  src/cli/persona_cmd.rs   281   `aizen persona …` / `aizen soul …`
  src/ui/plain_input.rs    256   the plain-mode chat box (`read_input_box`)
  src/cli/where_report.rs  240   `aizen where` / `/where`
  src/cli/skill_cmd.rs     237   `aizen skill …`
  src/ui/effort_ui.rs      135   the `/effort` slider + per-turn tier line

Two things made this cheap. Only eight main.rs functions were ever reached from another
module, and `src/tests/main_suite.rs` is a `#[path]` child of the crate root, so
`pub(crate) use crate::cli::…::*` at the root keeps both the dispatch in `main` and
`use super::*` in the tests resolving exactly as before. `src/ui/menus.rs` reaches these
names through `use crate::*`, which is why the re-exports are `pub(crate)` rather than
private imports.

`read_stdin` moved to `src/cli/mod.rs`: six subcommands share it, so it belongs beside
them rather than in the file they left.

Two documentation defects surfaced while cutting and are fixed here, since carrying them
into new files would have preserved them:

- the `/// /memory <sub> …` doc block sat on `sessions_menu`, describing something else
  entirely — it was orphaned when an earlier pass moved `slash_memory` to
  `features/slash_handlers.rs`. Reattached there; `sessions_menu` got an accurate doc.
- the `time machine (git snapshots)` banner sat above `run_team`/`run_work`, which are the
  multi-session coop commands. Dropped; its content is now `cli/time.rs`'s module doc.

Note for the next extraction: `#[path = "tests/rel_time.rs"]` has to become
`../tests/rel_time.rs` once its owner lives in a subdirectory.

Verified: cargo check --all-targets clean after every individual cut; 1635 passed / 0
failed / 2 ignored (the pre-move count, so nothing dropped out of the build);
cargo fmt --all -- --check clean; clippy 125 warnings, byte-identical to the baseline.
