+++ src/main.rs
@@ -21,25 +21,35 @@ mod skills; // skill store + registry
 mod ui; // tui · theme · markdown · spinner · splash · icons · image_input
 
 // Declaration-only: the command-line surface (clap derive types).
-mod cli_args;
+mod cli;
+mod cli_args; // one module per `aizen <subcommand>` — the behaviour behind `cli_args`
 
 // The reorg moved 23 top-level files into the folders above. These re-exports keep the
 // call sites in THIS file referring to the modules by their short names (no behavior
 // change) — every other file already uses the new `crate::<group>::<mod>` paths.
-use crate::agent::app_catalog;
 use crate::agent::prompt_lanes::*;
+// Subcommand bodies. Glob-imported so the dispatch in `main` keeps calling them by bare name, and
+// so `src/tests/main_suite.rs` (a `#[path]` child of the crate root) still resolves them through
+// `use super::*` — the same arrangement the earlier extractions use. `pub(crate)` because these
+// names USED to live at the crate root, and `src/ui/menus.rs` reaches them through `use crate::*`;
+// re-exporting keeps that reachability byte-identical to before the move.
+pub(crate) use crate::cli::read_stdin;
+pub(crate) use crate::cli::{
+    agents_cmd::*, apps::*, persona_cmd::*, sessions::*, skill_cmd::*, time::*, where_report::*,
+};
 use crate::core::session_store::*;
 use crate::core::{cli_config, config, types};
 use crate::features::slash_handlers::{
     handle_slash, slash_is_interactive, slash_menu, team_status_lines, SlashOutcome,
 };
-use crate::features::{commands, coop, crawl, cron, timemachine};
+use crate::features::{commands, coop, crawl, cron};
 use crate::hostbot::platforms::telegram;
 use crate::llm::client;
-use crate::persona::soul;
 use crate::skills::{self as skill, registry as skill_registry};
 use crate::ui::context_report::*;
-use crate::ui::menus::{apps_menu, persona_self_view_n, run_discord, run_telegram};
+pub(crate) use crate::ui::effort_ui::*;
+use crate::ui::menus::{apps_menu, run_discord, run_telegram};
+pub(crate) use crate::ui::plain_input::read_input_box;
 use crate::ui::{config_ui, icons, image_input, splash, theme, tui};
 
 // The `clap` type tree (every subcommand enum) lives in its own file — see `cli_args`.
@@ -47,10 +57,10 @@ use cli_args::*;
 
 use crate::core::approval::ApprovalMode;
 use agent::{AgentConfig, AgentOutcome, StopReason};
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{bail, Context, Result};
 use clap::Parser;
 use console::{style, Style};
-use dialoguer::{theme::ColorfulTheme, Confirm, Input, Password, Select};
+use dialoguer::{theme::ColorfulTheme, Confirm, Select};
 use types::{Message, ToolDef};
 
 /// Suppress Windows "hard error" dialogs process-wide (and for every child we spawn, which inherits
@@ -314,485 +324,6 @@ fn run_prompt_size(model: Option<String>, show_tools: bool, as_json: bool) -> Re
     Ok(())
 }
 
-/// `aizen apps …` — connect apps via the MCP registry.
-async fn run_apps(cmd: Option<AppsCmd>) -> Result<()> {
-    match cmd {
-        None | Some(AppsCmd::List) => {
-            apps_print_list();
-            Ok(())
-        }
-        Some(AppsCmd::Search { query, limit }) => {
-            let q = query.join(" ");
-            if q.trim().is_empty() {
-                return Err(anyhow!("usage: aizen apps search <keywords>"));
-            }
-            let hits = app_catalog::dedupe_latest(
-                app_catalog::search(q.trim(), limit.unwrap_or(20).clamp(1, 50)).await?,
-            );
-            if hits.is_empty() {
-                println!("no apps on {} match '{q}'", app_catalog::registry_base());
-                return Ok(());
-            }
-            println!(
-                "{}",
-                style(format!(
-                    "{} result(s) from {} — `aizen apps add <name>` to connect:",
-                    hits.len(),
-                    app_catalog::registry_base()
-                ))
-                .dim()
-            );
-            for s in &hits {
-                let name = style(&s.name).color256(splash::ACCENT);
-                println!("  {name}\n    {}", s.summary_line());
-            }
-            Ok(())
-        }
-        Some(AppsCmd::Add { name }) => apps_add(&name).await,
-        Some(AppsCmd::Info { name }) => apps_info(&name).await,
-        Some(AppsCmd::Login { name }) => {
-            crate::agent::mcp::login(&name).await?;
-            println!(
-                "{}",
-                style(format!("✓ signed in to '{name}'. Its tools load on your next message (/mcp to verify)."))
-                    .color256(splash::ACCENT)
-            );
-            Ok(())
-        }
-        Some(AppsCmd::Remove { name }) => {
-            if app_catalog::remove_server(&name)? {
-                crate::agent::mcp_oauth::clear_token(&name); // drop any cached OAuth token too
-                crate::agent::mcp::invalidate();
-                println!(
-                    "{}",
-                    style(format!("✓ disconnected '{name}'.")).color256(splash::ACCENT)
-                );
-            } else {
-                println!("no connected app keyed '{name}' (see `aizen apps list`).");
-            }
-            Ok(())
-        }
-    }
-}
-
-/// Render the featured catalog with connection badges.
-fn apps_print_list() {
-    let installed = app_catalog::installed_keys();
-    println!(
-        "{}",
-        style("Apps — connect via the MCP registry (`aizen apps add <key>`):").bold()
-    );
-    for f in app_catalog::FEATURED {
-        let on = installed.iter().any(|k| k == f.key);
-        let badge = if on {
-            style("✓").color256(splash::ACCENT).to_string()
-        } else {
-            style("○").dim().to_string()
-        };
-        println!(
-            "  {badge}  {} {:<18} {}",
-            icons::g(f.icon),
-            style(f.key).color256(splash::ACCENT),
-            style(f.blurb).dim()
-        );
-    }
-    // Apps the user connected that aren't in the featured set (added via `aizen apps add <name>`).
-    let custom: Vec<&String> = installed
-        .iter()
-        .filter(|k| !app_catalog::FEATURED.iter().any(|f| f.key == **k))
-        .collect();
-    if !custom.is_empty() {
-        println!("\n{}", style("connected (custom):").bold());
-        for k in &custom {
-            println!(
-                "  {}  {} {}",
-                style("✓").color256(splash::ACCENT),
-                icons::g("🧩"),
-                style(k).color256(splash::ACCENT)
-            );
-        }
-    }
-    println!(
-        "\n{}",
-        style("details: `aizen apps info <key>`   ·   search: `aizen apps search <keywords>`   ·   remove: `aizen apps remove <key>`").dim()
-    );
-}
-
-/// Resolve a featured key or registry name → fetch spec → pick transport → prompt secrets → write
-/// the mcp.json entry. Interactive (hidden secret prompts); transparent about what it chose.
-pub(crate) async fn apps_add(name: &str) -> Result<()> {
-    let theme = ui_theme();
-    // Resolve to the VIABLE candidate set and let the user CHOOSE (publisher + local/hosted shown) —
-    // connecting an app hands it your token, so we never silently wire whatever sorts first. The
-    // best heuristic match (pick_best) is the pre-selected default. (A featured app's vendor is just
-    // a search hint + default; the official server is often OAuth-only, so community servers appear.)
-    let (key0, query, prefer, label) = match app_catalog::featured(name) {
-        Some(f) => (
-            Some(f.key.to_string()),
-            f.query.to_string(),
-            f.prefer.to_string(),
-            f.label.to_string(),
-        ),
-        None => (None, name.to_string(), name.to_string(), name.to_string()),
-    };
-    let hits = app_catalog::dedupe_latest(app_catalog::search(&query, 50).await?);
-    let viable: Vec<app_catalog::RegistryServer> = hits
-        .into_iter()
-        .filter(|s| app_catalog::is_viable(s))
-        .collect();
-    if viable.is_empty() {
-        return Err(anyhow!(
-            "no connectable '{label}' server found on the registry (only legacy sse-only entries, which aizen's client doesn't speak). Run `aizen apps search {query}` to explore."
-        ));
-    }
-    let default_idx = app_catalog::pick_best(&viable, &prefer)
-        .and_then(|best| viable.iter().position(|s| s.name == best.name))
-        .unwrap_or(0);
-    // One clean, COLUMN-ALIGNED line per server: ★ recommended · transport · short name · short desc.
-    // (The old 2-line label repeated the name and let long descriptions wrap into a wall of text.)
-    let trunc = |s: &str, n: usize| -> String {
-        let s = s.replace(['\n', '\r'], " ");
-        if s.chars().count() <= n {
-            s
-        } else {
-            format!(
-                "{}…",
-                s.chars().take(n.saturating_sub(1)).collect::<String>()
-            )
-        }
-    };
-    let name_w = viable
-        .iter()
-        .map(|s| s.short_name().chars().count())
-        .max()
-        .unwrap_or(8)
-        .clamp(8, 30);
-    let tag_w = viable
-        .iter()
-        .map(|s| s.transport_tag().chars().count())
-        .max()
-        .unwrap_or(7)
-        .clamp(7, 14);
-    let mut labels: Vec<String> = viable
-        .iter()
-        .enumerate()
-        .map(|(i, s)| {
-            let star = if i == default_idx { "★" } else { " " };
-            let tag = format!("{:<w$}", trunc(&s.transport_tag(), tag_w), w = tag_w);
-            let nm = format!("{:<w$}", trunc(&s.short_name(), name_w), w = name_w);
-            format!("{star} {tag}  {nm}  {}", trunc(&s.description, 52))
-        })
-        .collect();
-    labels.push("Cancel".to_string());
-    println!(
-        "{}",
-        style(format!("Connect {label}  —  ★ recommended · local·X = your machine · sign-in = OAuth · hosted = third party")).dim()
-    );
-    let idx = match Select::with_theme(&theme)
-        .with_prompt("Pick a server (↑↓, Enter)")
-        .items(&labels)
-        .default(default_idx)
-        .interact_opt()?
-    {
-        Some(i) if i < viable.len() => i,
-        _ => {
-            println!("{}", style("cancelled.").dim());
-            return Ok(());
-        }
-    };
-    let server = viable[idx].clone();
-    let key = key0.unwrap_or_else(|| app_catalog::slug_from_name(&server.name));
-
-    // Runtime-aware: prefer a transport whose runner is actually on PATH (don't wire an npx server
-    // when Node isn't installed if a remote would work).
-    let choice = app_catalog::pick_transport_for_install(&server)
-        .context("this server declares no transport aizen can use")?;
-    let repo = server
-        .repository
-        .as_ref()
-        .map(|r| r.url.clone())
-        .unwrap_or_default();
-    println!(
-        "{}",
-        style(format!("→ {}", server.name)).color256(splash::ACCENT)
-    );
-    if !server.description.is_empty() {
-        println!("  {}", style(&server.description).dim());
-    }
-    if !repo.is_empty() {
-        println!("  {}", style(&repo).dim());
-    }
-    if let Some(rt) = app_catalog::runtime_prereq(&server, choice) {
-        let have = which_runtime(rt);
-        let note = if have {
-            format!("runs locally via {rt} (found)")
-        } else {
-            format!("runs locally via {rt} — NOT found on PATH; install it to run this app")
-        };
-        println!("  {}", style(note).dim());
-    }
-    // Static-token hosted remote → an explicit host-named confirm before we collect a token (it
-    // leaves your machine for a third party). This is the strongest gate; refusing aborts the connect.
-    if let app_catalog::TransportChoice::Remote(i) = choice {
-        let host = server
-            .remotes
-            .get(i)
-            .map(|r| app_catalog::host_of(&r.url))
-            .unwrap_or_default();
-        println!(
-            "  {}",
-            style(format!(
-                "⚠ hosted remote @ {host} — a third party runs this server."
-            ))
-            .yellow()
-        );
-        let go = Confirm::with_theme(&theme)
-            .with_prompt(format!(
-                "Send your credentials to '{host}' (a third party)?"
-            ))
-            .default(false)
-            .interact()
-            .unwrap_or(false);
-        if !go {
-            println!(
-                "{}",
-                style("cancelled — no third-party remote connected.").dim()
-            );
-            return Ok(());
-        }
-    }
-    // OAuth remote → you authenticate directly with the vendor (no token leaves via us); confirm we
-    // may open the browser to sign in.
-    if let app_catalog::TransportChoice::OAuthRemote(i) = choice {
-        let host = server
-            .remotes
-            .get(i)
-            .map(|r| app_catalog::host_of(&r.url))
-            .unwrap_or_default();
-        println!(
-            "  {}",
-            style(format!(
-                "🔐 sign-in app @ {host} — Aizen will open your browser to authorize."
-            ))
-            .dim()
-        );
-        let go = Confirm::with_theme(&theme)
-            .with_prompt(format!("Connect '{host}' and sign in now?"))
-            .default(true)
-            .interact()
-            .unwrap_or(false);
-        if !go {
-            println!("{}", style("cancelled.").dim());
-            return Ok(());
-        }
-    }
-
-    // Collect any declared secrets (hidden), with a confirm gate (we're writing a token to disk).
-    let mut ask = |spec: &app_catalog::PromptSpec| -> String {
-        let prompt = if spec.description.is_empty() {
-            format!("{} ", spec.label)
-        } else {
-            format!("{} ({})", spec.label, spec.description)
-        };
-        let val = if spec.is_secret {
-            Password::with_theme(&theme)
-                .with_prompt(prompt.trim())
-                .allow_empty_password(true)
-                .interact()
-                .unwrap_or_default()
-        } else {
-            Input::<String>::with_theme(&theme)
-                .with_prompt(prompt.trim())
-                .allow_empty(true)
-                .interact_text()
-                .unwrap_or_default()
-        };
-        val.trim().to_string()
-    };
-    let entry = app_catalog::build_entry(&server, choice, &mut ask)?;
-
-    // Confirm gate (we're about to write a token to disk) — show the resolved entry with secrets
-    // MASKED so the user sees exactly what gets written before committing.
-    println!("\n{}", style(format!("About to connect '{key}':")).bold());
-    print_entry_summary(&entry, Some(&key));
-    let ok = Confirm::with_theme(&theme)
-        .with_prompt("Write this to mcp.json?")
-        .default(true)
-        .interact()
-        .unwrap_or(false);
-    if !ok {
-        println!("{}", style("cancelled — nothing written.").dim());
-        return Ok(());
-    }
-    app_catalog::write_server(&key, entry)?;
-    crate::agent::mcp::invalidate(); // hot-reload: the next message reconnects from the new mcp.json
-
-    // OAuth app → run the browser sign-in right now so it's usable immediately. A failure isn't fatal:
-    // the entry is written, the user can retry with `aizen apps login <key>`.
-    if matches!(choice, app_catalog::TransportChoice::OAuthRemote(_)) {
-        match crate::agent::mcp::login(&key).await {
-            Ok(()) => println!(
-                "{}",
-                style(format!("✓ connected & signed in to '{key}'. Its tools load on your next message (/mcp to verify)."))
-                    .color256(splash::ACCENT)
-            ),
-            Err(e) => println!(
-                "{}",
-                style(format!("connected '{key}', but sign-in didn't finish — {e:#}\n  finish it with `aizen apps login {key}`."))
-                    .yellow()
-            ),
-        }
-        return Ok(());
-    }
-    println!(
-        "{}",
-        style(format!(
-            "✓ connected '{key}'.  Its tools load on your next message (/mcp to verify)."
-        ))
-        .color256(splash::ACCENT)
-    );
-    Ok(())
-}
-
-/// Print an mcp.json entry's transport + config with secret VALUES masked (shared by the add-confirm
-/// preview and `apps info`). Never prints a token value — presence only. `key` (when known) lets it
-/// show OAuth sign-in state from the token cache.
-fn print_entry_summary(entry: &serde_json::Value, key: Option<&str>) {
-    if let Some(url) = entry.get("url").and_then(|v| v.as_str()) {
-        println!("  {} remote (streamable-http)", style("transport").dim());
-        println!("  {} {url}", style("url      ").dim());
-        println!(
-            "  {} {}",
-            style("host     ").dim(),
-            style(app_catalog::host_of(url)).dim()
-        );
-        if entry.get("auth").and_then(|v| v.as_str()) == Some("oauth") {
-            let signed = key.map(crate::agent::mcp_oauth::has_token).unwrap_or(false);
-            let state = if signed {
-                "signed in".to_string()
-            } else {
-                "not signed in — `aizen apps login <key>`".to_string()
-            };
-            println!("  {} oauth ({state})", style("auth     ").dim());
-        }
-    } else if let Some(cmd) = entry.get("command").and_then(|v| v.as_str()) {
-        let args = entry
-            .get("args")
-            .and_then(|a| a.as_array())
-            .map(|a| {
-                a.iter()
-                    .filter_map(|x| x.as_str())
-                    .collect::<Vec<_>>()
-                    .join(" ")
-            })
-            .unwrap_or_default();
-        println!("  {} local (stdio)", style("transport").dim());
-        println!("  {} {cmd} {args}", style("command  ").dim());
-        if !cmd.contains(['/', '\\']) {
-            let have = which_runtime(cmd);
-            let note = if have {
-                format!("{cmd}: found on PATH")
-            } else {
-                format!("{cmd}: NOT on PATH — install it to run this app")
-            };
-            println!("  {} {note}", style("runtime  ").dim());
-        }
-    }
-    for field in ["env", "headers"] {
-        if let Some(obj) = entry.get(field).and_then(|v| v.as_object()) {
-            for (k, v) in obj {
-                println!(
-                    "  {} {k} = {}",
-                    style(format!("{field:<8}")).dim(),
-                    mask_secret(v.as_str().unwrap_or(""))
-                );
-            }
-        }
-    }
-}
-
-/// Mask a secret/config value for display: presence only, never the value (the standing key-safety
-/// rule). Empty → "(empty)"; set → "•••• (set)".
-fn mask_secret(v: &str) -> String {
-    if v.trim().is_empty() {
-        style("(empty)").dim().to_string()
-    } else {
-        style("•••• (set)").dim().to_string()
-    }
-}
-
-/// `aizen apps info <key>` — the detail view for ONE connected app: its mcp.json config (transport +
-/// secrets MASKED) plus a LIVE probe (handshake + the tools it actually exposes, or why it failed).
-pub(crate) async fn apps_info(key: &str) -> Result<()> {
-    let Some(entry) = app_catalog::installed_entry(key) else {
-        return Err(anyhow!(
-            "no connected app keyed '{key}' — see `aizen apps list`"
-        ));
-    };
-    println!("{}", style(key).color256(splash::ACCENT).bold());
-    print_entry_summary(&entry, Some(key));
-
-    // Live probe.
-    println!("  {}", style("probing (connect + tools/list)…").dim());
-    match crate::agent::mcp::probe(key).await {
-        Ok(rep) => {
-            let info = rep.server_info.get("serverInfo");
-            let sname = info
-                .and_then(|s| s.get("name"))
-                .and_then(|v| v.as_str())
-                .unwrap_or(key);
-            let sver = info
-                .and_then(|s| s.get("version"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("");
-            println!(
-                "  {} {}",
-                style("✓").color256(splash::ACCENT),
-                style(format!("{sname} {sver}  ·  {} tool(s)", rep.tools.len())).bold()
-            );
-            for t in &rep.tools {
-                let ro = if t.read_only {
-                    style(" [read-only]").dim().to_string()
-                } else {
-                    String::new()
-                };
-                let d: String = t.description.chars().take(72).collect();
-                println!(
-                    "    {}{ro}  {}",
-                    style(&t.name).color256(splash::ACCENT),
-                    style(d).dim()
-                );
-            }
-            if rep.tools.is_empty() {
-                println!("    {}", style("(this server advertised no tools)").dim());
-            }
-        }
-        // `{e:#}` = the full anyhow chain (includes the server's stderr tail captured by the client).
-        Err(e) => println!("  {}", style(format!("✗ could not connect — {e:#}")).red()),
-    }
-    Ok(())
-}
-
-/// Best-effort PATH check for a runner (npx/uvx/docker) — Windows adds `.cmd`/`.exe` variants.
-fn which_runtime(rt: &str) -> bool {
-    let exts: &[&str] = if cfg!(windows) {
-        &["", ".cmd", ".exe", ".bat"]
-    } else {
-        &[""]
-    };
-    let Some(path) = std::env::var_os("PATH") else {
-        return false;
-    };
-    for dir in std::env::split_paths(&path) {
-        for ext in exts {
-            if dir.join(format!("{rt}{ext}")).is_file() {
-                return true;
-            }
-        }
-    }
-    false
-}
-
 async fn run_crawl(args: CrawlArgs) -> Result<()> {
     let opts = crawl::CrawlOptions {
         seeds: args.urls,
@@ -922,238 +453,6 @@ async fn run_agent_capture(
         .unwrap_or_else(|| "(the agent produced no answer)".to_string()))
 }
 
-/// Strip URL userinfo before display when it carries a password/token
-/// (`https://user:TOKEN@host/…`) — remote URLs may embed credentials and the identity surfaces
-/// must never print one. A plain username (`git@host:…`) is kept: it isn't a secret and losing
-/// it would make the URL unrecognizable.
-fn redact_remote_url(url: &str) -> String {
-    let (scheme, rest) = match url.find("://") {
-        Some(i) => url.split_at(i + 3),
-        None => ("", url),
-    };
-    match rest.find('@') {
-        Some(at) if rest[..at].contains(':') => format!("{scheme}***@{}", &rest[at + 1..]),
-        _ => url.to_string(),
-    }
-}
-
-/// How many `*.md` files a store directory holds, and a `(not created yet)` note when it doesn't
-/// exist. Shared by both `where` reports so an absent folder never reads as an empty one.
-fn dir_count_line(label: &str, p: &std::path::Path, unit: &str) -> String {
-    if !p.exists() {
-        return format!("  {label:<8}: {}   (not created yet)", p.display());
-    }
-    let n = std::fs::read_dir(p)
-        .map(|rd| {
-            rd.flatten()
-                .filter(|e| e.path().extension().and_then(|x| x.to_str()) == Some("md"))
-                .count()
-        })
-        .unwrap_or(0);
-    format!("  {label:<8}: {}   {n} {unit}", p.display())
-}
-
-/// Where the memory store physically lives, per directory, with counts.
-///
-/// `memory list` names three commands and every one of them edits a SINGLE entry by id. Bulk work —
-/// deleting forty near-duplicates, fixing a wrong word across many facts — is a file-manager job, and
-/// until now the only place any path appeared was `memory show <id>`'s `file:` line, one entry at a
-/// time. Naming the review dir matters most: 29 queued candidates sat there unreadable because
-/// nothing said they were on disk at all.
-fn memory_where_report() -> String {
-    use std::fmt::Write as _;
-    let mut s = String::new();
-    let _ = writeln!(
-        s,
-        "{}",
-        dir_count_line("entries", &crate::core::config::entries_dir(), "fact(s)")
-    );
-    let _ = writeln!(
-        s,
-        "{}",
-        dir_count_line("review", &crate::core::config::review_dir(), "awaiting")
-    );
-    let _ = writeln!(
-        s,
-        "{}",
-        dir_count_line("archive", &crate::core::config::archive_dir(), "retired")
-    );
-    let graph = crate::core::config::graph_path();
-    let edges = std::fs::read_to_string(&graph)
-        .map(|r| r.lines().filter(|l| !l.trim().is_empty()).count())
-        .unwrap_or(0);
-    let _ = writeln!(s, "  {:<8}: {}   {edges} edge(s)", "graph", graph.display());
-    let _ = writeln!(
-        s,
-        "  {:<8}: {}",
-        "core",
-        crate::core::config::style_path().display()
-    );
-    let _ = write!(
-        s,
-        "{}",
-        style(
-            "\nEdit or delete files directly — they are plain markdown with a frontmatter header.\n\
-             Re-run `aizen memory doctor` afterwards to catch anything left dangling."
-        )
-        .dim()
-    );
-    s
-}
-
-/// Where skills are read from — all three roots, because `skill list`'s `[project]`/`[repo]` tags
-/// say which root a row came from without saying where that root is, and auto-learned skills land in
-/// the zone dir whose slug (`p/admin-5296147b`) is not guessable from the project name.
-fn skill_where_report() -> String {
-    use std::fmt::Write as _;
-    let mut s = String::new();
-    let _ = writeln!(
-        s,
-        "{}",
-        dir_count_line("global", &crate::skills::skills_dir(), "skill(s)")
-    );
-    let _ = writeln!(
-        s,
-        "{}",
-        dir_count_line("zone", &crate::skills::project_zone_dir(), "skill(s)")
-    );
-    let _ = writeln!(
-        s,
-        "  {}",
-        style(format!(
-            "         ↑ auto-learned skills for zone {}",
-            crate::core::config::project_slug()
-        ))
-        .dim()
-    );
-    let _ = write!(
-        s,
-        "{}",
-        dir_count_line("repo", &crate::skills::project_skills_dir(), "skill(s)")
-    );
-    s
-}
-
-/// The identity card — one honest surface for the questions that previously had none: which
-/// root am I in, which zone does my memory go to, which git binary runs, where do sessions live.
-/// Shared verbatim by `aizen where` (println) and `/where` (tui::emit_line).
-pub(crate) fn where_report() -> String {
-    use std::fmt::Write as _;
-    let root = crate::core::config::project_root();
-    let slug = crate::core::config::project_slug();
-    let home = crate::core::config::aizen_home();
-    let cwd = std::env::current_dir()
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "?".into());
-    let mut s = String::new();
-    let _ = writeln!(s, "project root : {}", root.display());
-    if let Ok(over) = std::env::var("AIZEN_PROJECT_ROOT") {
-        if !over.trim().is_empty() {
-            let _ = writeln!(
-                s,
-                "               (root forced by AIZEN_PROJECT_ROOT={})",
-                over.trim()
-            );
-        }
-    }
-    let _ = writeln!(
-        s,
-        "cwd          : {cwd}   (identity follows the root, fixed at launch)"
-    );
-    let _ = writeln!(
-        s,
-        "zone slug    : {slug}   (keys memory scope · skills · codebase index · frozen core)"
-    );
-    if let Some(url) = crate::core::config::git_remote_origin(&root) {
-        let _ = writeln!(
-            s,
-            "git remote   : {}   (informational — no longer part of the identity key)",
-            redact_remote_url(&url)
-        );
-    }
-    match crate::core::gitx::git_exe() {
-        Some(p) => {
-            let _ = writeln!(s, "git          : {}", p.display());
-        }
-        None => {
-            let _ = writeln!(
-                s,
-                "git          : NOT FOUND — identity uses the nearest .git marker (or this folder); time-machine checkpoints are off"
-            );
-        }
-    }
-    if let Some(note) = crate::core::gitx::resolution_note() {
-        if crate::core::gitx::git_exe().is_some() {
-            let _ = writeln!(s, "               ({note})");
-        }
-    }
-    let zone_dir = crate::skills::project_zone_dir();
-    let idx = crate::core::config::codebase_index_path(&slug);
-    let exists = |p: &std::path::Path| {
-        if p.exists() {
-            ""
-        } else {
-            "   (not created yet)"
-        }
-    };
-    let _ = writeln!(s, "home         : {}", home.display());
-    let _ = writeln!(
-        s,
-        "memory store : {}",
-        crate::core::config::cli_memory_dir().display()
-    );
-    let _ = writeln!(
-        s,
-        "skills zone  : {}{}",
-        zone_dir.display(),
-        exists(&zone_dir)
-    );
-    let _ = writeln!(s, "codebase idx : {}{}", idx.display(), exists(&idx));
-    let _ = writeln!(s, "sessions     : {}", sessions_dir().display());
-    if let Some(n) = sessions_with_secrets() {
-        let _ = writeln!(
-            s,
-            "⚠ secrets     : {n} saved transcript(s) contain credential-shaped text — a key pasted into a chat is stored verbatim. Open the folder above and edit or delete them."
-        );
-    }
-    if let Some(l) = crate::features::zones::quick_legacy_probe() {
-        let _ = writeln!(
-            s,
-            "⚠ legacy zone : {l} — data from the old slug keying; `aizen zone migrate` shows what would merge (--apply to do it)"
-        );
-    }
-    s.trim_end().to_string()
-}
-
-/// How many saved transcripts hold credential-shaped text, or `None` when none do.
-///
-/// Names are guarded at derivation (see [`suggest_session_name`]), but a key pasted into a chat is
-/// still in that file's message text: a saved session is a verbatim transcript, and nothing redacts
-/// it on the way to disk. Deleting or rewriting a user's own conversation history is not a call this
-/// tool makes on its own, so `/where` reports the count and names the folder — the number is
-/// actionable, and the values are never printed.
-///
-/// Uses the vendor-prefix test, NOT the shape test that guards name derivation. Measured on the 27
-/// real transcripts here: prefix matches 12 strings, all real keys; shape matched 5170, of which 4026
-/// were ISO timestamps (long, mixed-case, letters and digits — indistinguishable from key material by
-/// shape alone). A warning that fires on every file teaches the user to ignore it.
-///
-/// Counts FILES, not occurrences: the useful signal is "which files do I need to open".
-fn sessions_with_secrets() -> Option<usize> {
-    let rd = std::fs::read_dir(sessions_dir()).ok()?;
-    let n = rd
-        .flatten()
-        .filter(|e| e.path().extension().and_then(|x| x.to_str()) == Some("json"))
-        .filter(|e| {
-            std::fs::read_to_string(e.path()).is_ok_and(|raw| {
-                raw.split(|c: char| c.is_whitespace() || matches!(c, '"' | ',' | '\\'))
-                    .any(crate::core::slug::has_vendor_key_prefix)
-            })
-        })
-        .count();
-    (n > 0).then_some(n)
-}
-
 fn run_zone(cmd: ZoneCmd) -> Result<()> {
     match cmd {
         ZoneCmd::Migrate { apply } => {
@@ -1192,8 +491,6 @@ fn run_zone(cmd: ZoneCmd) -> Result<()> {
         }
     }
 }
-
-// ───────────────────────────── time machine (git snapshots) ─────────────────────────────
 
 /// `aizen team …` — the non-interactive twin of `/team`. Same registry, same plan, `println!` instead
 /// of `tui::emit_line` (no retained frame is up here, so a raw print is the correct surface).
@@ -1365,580 +662,6 @@ fn run_work(cmd: WorkCmd) -> Result<()> {
     }
 }
 
-fn run_time(cmd: TimeCmd) -> Result<()> {
-    match cmd {
-        TimeCmd::Save { label } => {
-            let snap = timemachine::save(&label.join(" "), false)?;
-            println!(
-                "{} #{}  {}",
-                style("✓ checkpoint").color256(splash::ACCENT),
-                snap.id,
-                style(&snap.created).dim()
-            );
-            Ok(())
-        }
-        TimeCmd::List => {
-            print_timeline()?;
-            Ok(())
-        }
-        TimeCmd::Restore { id } => {
-            let snap = timemachine::restore(id)?;
-            let label = if snap.label.is_empty() {
-                "(no label)".to_string()
-            } else {
-                snap.label.clone()
-            };
-            println!(
-                "{} #{} — {label}",
-                style("⏪ restored to").color256(splash::ACCENT),
-                snap.id
-            );
-            // Say WHAT changed and that it's undoable: aizen only rewinds the working tree (files),
-            // never your chat/history — and because the pre-restore state was auto-snapshotted, you
-            // can always go forward again (`aizen time redo`, or restore the newest checkpoint).
-            println!("{}", style("  files only — your conversation is untouched · reversible with `aizen time redo`").dim());
-            Ok(())
-        }
-        TimeCmd::Diff {
-            from,
-            to,
-            patch,
-            paths,
-            json,
-        } => run_time_diff(from, to, paths, patch, json),
-        TimeCmd::Undo => {
-            let snap = timemachine::undo()?;
-            println!(
-                "{} #{}",
-                style("⏪ undo →").color256(splash::ACCENT),
-                snap.id
-            );
-            Ok(())
-        }
-        TimeCmd::Redo => {
-            let snap = timemachine::redo()?;
-            println!(
-                "{} #{}",
-                style("⏩ redo →").color256(splash::ACCENT),
-                snap.id
-            );
-            Ok(())
-        }
-        TimeCmd::Prune { keep } => {
-            let k = keep.or(cli_config::load().timemachine_keep).unwrap_or(50);
-            let dropped = timemachine::prune(k)?;
-            println!(
-                "{} {dropped} old checkpoint(s); kept ≤{k}.",
-                style("🧹 pruned").color256(splash::ACCENT)
-            );
-            Ok(())
-        }
-        TimeCmd::Doctor { json, repair } => {
-            let report = if repair {
-                timemachine::doctor_repair()?
-            } else {
-                timemachine::doctor()?
-            };
-            if json {
-                println!("{}", serde_json::to_string_pretty(&report)?);
-            } else {
-                println!(
-                    "{}  repo {} · worktree {} · {} checkpoint(s)",
-                    if report.ok {
-                        "✓ time machine healthy"
-                    } else {
-                        "⚠ time machine needs attention"
-                    },
-                    report.repo_id,
-                    report.worktree_id,
-                    report.checkpoints
-                );
-                println!("  store {}", report.store);
-                // Loose objects are not an "issue" — a store with 15k of them still restores fine.
-                // They are a disk-growth signal, so surface them only once there are enough to matter
-                // and say what fixes it, rather than printing a zero on every healthy run.
-                if report.loose_objects >= 1024 {
-                    println!(
-                        "  {} loose object(s) not packed — run `aizen time gc` to compact",
-                        report.loose_objects
-                    );
-                }
-                for issue in &report.issues {
-                    println!("  - {issue}");
-                }
-            }
-            if !report.ok {
-                bail!("time-machine doctor found {} issue(s)", report.issues.len());
-            }
-            Ok(())
-        }
-        TimeCmd::Gc { all, apply } => {
-            if all {
-                let report = timemachine::gc_all(apply)?;
-                let fmt_mb = |b: u64| format!("{:.1} MB", b as f64 / 1_048_576.0);
-                if report.orphans.is_empty() {
-                    println!(
-                        "{} {} store(s) scanned · no orphans (every source repo still exists)",
-                        style("🧹 time gc --all:").color256(splash::ACCENT),
-                        report.stores.len()
-                    );
-                } else {
-                    let total: u64 = report.orphans.iter().map(|o| o.bytes).sum();
-                    println!(
-                        "{} {} orphan store(s), {} reclaimable:",
-                        style("🧹 time gc --all:").color256(splash::ACCENT),
-                        report.orphans.len(),
-                        fmt_mb(total)
-                    );
-                    for o in &report.orphans {
-                        println!(
-                            "  {} · {} · {} checkpoint(s) · source gone: {}",
-                            o.repo_id,
-                            fmt_mb(o.bytes),
-                            o.checkpoints,
-                            o.source.as_deref().unwrap_or("(unknown)")
-                        );
-                    }
-                    if report.applied {
-                        // `--apply` RENAMES; nothing here or anywhere else ever empties the trash.
-                        // Saying only "moved" alongside a "reclaimable" figure reads as though the
-                        // disk came back, and it has not — so name the step that actually frees it.
-                        let trash = report.trash_dir.as_deref().unwrap_or("(trash)");
-                        println!("  → moved to {trash}");
-                        println!(
-                            "  {} still on disk until you delete that directory — nothing reaps it for you",
-                            style("note:").bold()
-                        );
-                    } else {
-                        println!(
-                            "  (dry-run — re-run with {} to move these to .trash/)",
-                            style("--apply").bold()
-                        );
-                    }
-                }
-                Ok(())
-            } else {
-                let (report, compacted) = timemachine::doctor_gc()?;
-                println!(
-                    "{} repo {} · worktree {} · {} checkpoint(s)",
-                    style("🧹 time metadata cleaned:").color256(splash::ACCENT),
-                    report.repo_id,
-                    report.worktree_id,
-                    report.checkpoints
-                );
-                // Report bytes actually returned, not objects touched: packing 14k objects sounds
-                // like work, but the number the user came for is how much smaller the store got.
-                if let Some(c) = compacted.filter(|c| c.packed > 0) {
-                    // Scale the unit: a small store printed as "0.0 MB → 0.0 MB" reads as a no-op
-                    // when 36 objects were in fact packed.
-                    let size = |b: u64| {
-                        if b >= 1_048_576 {
-                            format!("{:.1} MB", b as f64 / 1_048_576.0)
-                        } else {
-                            format!("{:.0} KB", b as f64 / 1024.0)
-                        }
-                    };
-                    println!(
-                        "  packed {} loose object(s) · {} → {}",
-                        c.packed,
-                        size(c.before_bytes),
-                        size(c.after_bytes)
-                    );
-                }
-                Ok(())
-            }
-        }
-        TimeCmd::Clear => {
-            let n = timemachine::clear()?;
-            println!(
-                "{} {n} checkpoint(s) deleted.",
-                style("🧹 cleared").color256(splash::ACCENT)
-            );
-            Ok(())
-        }
-    }
-}
-
-/// Resolve the `[FROM] [TO]` positional pair into two timeline sides.
-///
-/// The defaults encode the question people actually ask. Bare `time diff` means "what have I changed
-/// since the last checkpoint" (cursor → working tree), which is the state you want before deciding
-/// whether to keep or rewind. One argument means "since THAT point" (given → working tree), because
-/// naming a single checkpoint and getting a checkpoint↔checkpoint diff against an unnamed second
-/// point would be guesswork.
-fn resolve_diff_sides(
-    from: Option<&str>,
-    to: Option<&str>,
-) -> Result<(timemachine::DiffSide, timemachine::DiffSide)> {
-    use timemachine::DiffSide;
-    let parse = |s: &str| {
-        DiffSide::parse(s).with_context(|| {
-            format!("`{s}` is not a checkpoint id or `working` (try `aizen time list`)")
-        })
-    };
-    match (from, to) {
-        (None, _) => {
-            let (snaps, cursor) = timemachine::timeline()?;
-            let cur = cursor.and_then(|i| snaps.get(i)).map(|s| s.id).context(
-                "no checkpoints yet — nothing to diff against (`aizen time save` first)",
-            )?;
-            Ok((DiffSide::Checkpoint(cur), DiffSide::Working))
-        }
-        (Some(f), None) => Ok((parse(f)?, DiffSide::Working)),
-        (Some(f), Some(t)) => Ok((parse(f)?, parse(t)?)),
-    }
-}
-
-/// Render a diff report as display lines. Shared so `aizen time diff` (stdout) and `/diff` (the TUI,
-/// which MUST go through `tui::emit_line` or the render thread wipes the output) format identically.
-/// `narrow_hint` differs between the two because the flag spelling does: `--path p` vs `-- p`.
-pub(crate) fn diff_lines(report: &timemachine::DiffReport, narrow_hint: &str) -> Vec<String> {
-    if report.is_empty() {
-        return vec![style(format!(
-            "⎇ no changes between {} and {}",
-            report.from, report.to
-        ))
-        .dim()
-        .to_string()];
-    }
-    let mut out = vec![format!(
-        "{}  {} → {}  ·  {} file(s), {}",
-        style("⎇ diff").color256(splash::ACCENT).bold(),
-        report.from,
-        report.to,
-        report.files.len(),
-        style(format!(
-            "+{} -{}",
-            report.total_added(),
-            report.total_deleted()
-        ))
-        .dim(),
-    )];
-    for f in &report.files {
-        // `None` counts mean git reported `-`: a binary file, not a zero-line change.
-        let churn = match (f.added, f.deleted) {
-            (Some(a), Some(d)) => format!("+{a} -{d}"),
-            _ => "binary".to_string(),
-        };
-        let path = match &f.old_path {
-            Some(old) => format!("{old} → {}", f.path),
-            None => f.path.clone(),
-        };
-        out.push(format!("  {} {path}  {}", f.status, style(churn).dim()));
-    }
-    match &report.patch {
-        Some(text) => {
-            out.push(String::new());
-            out.extend(text.lines().map(|l| l.to_string()));
-            if report.patch_truncated {
-                out.push(
-                    style(format!("… patch truncated — narrow it with {narrow_hint}"))
-                        .dim()
-                        .to_string(),
-                );
-            }
-        }
-        None => out.push(
-            style(format!(
-                "  --patch for the full text · {narrow_hint} to narrow it"
-            ))
-            .dim()
-            .to_string(),
-        ),
-    }
-    out
-}
-
-/// `aizen time diff` — print the changes between two points in the timeline.
-fn run_time_diff(
-    from: Option<String>,
-    to: Option<String>,
-    paths: Vec<String>,
-    patch: bool,
-    json: bool,
-) -> Result<()> {
-    let report = build_time_diff(from, to, paths, patch)?;
-    if json {
-        println!("{}", serde_json::to_string_pretty(&report)?);
-        return Ok(());
-    }
-    for line in diff_lines(&report, "--path <p>") {
-        println!("{line}");
-    }
-    Ok(())
-}
-
-/// Cap on emitted patch bytes. Generous enough for a real review, bounded so a huge rewrite cannot
-/// flood the terminal (or, for the agent tool, the tool-result budget).
-const DIFF_PATCH_LIMIT: usize = 400 * 1024;
-
-pub(crate) fn build_time_diff(
-    from: Option<String>,
-    to: Option<String>,
-    paths: Vec<String>,
-    patch: bool,
-) -> Result<timemachine::DiffReport> {
-    let (a, b) = resolve_diff_sides(from.as_deref(), to.as_deref())?;
-    timemachine::diff(&a, &b, &paths, patch.then_some(DIFF_PATCH_LIMIT))
-}
-
-/// Human "2m ago" from a snapshot's stored LOCAL timestamp. Pure core (`rel_time_from`) takes `now`
-/// so the bucketing is unit-testable; a malformed timestamp degrades to the raw string.
-fn rel_time(created: &str) -> String {
-    rel_time_from(created, chrono::Local::now().naive_local())
-}
-fn rel_time_from(created: &str, now: chrono::NaiveDateTime) -> String {
-    match chrono::NaiveDateTime::parse_from_str(created, "%Y-%m-%d %H:%M:%S") {
-        Ok(t) => {
-            let secs = (now - t).num_seconds();
-            if secs < 0 {
-                "just now".to_string()
-            } else if secs < 60 {
-                format!("{secs}s ago")
-            } else if secs < 3600 {
-                format!("{}m ago", secs / 60)
-            } else if secs < 86_400 {
-                format!("{}h ago", secs / 3600)
-            } else {
-                format!("{}d ago", secs / 86_400)
-            }
-        }
-        Err(_) => created.to_string(),
-    }
-}
-
-#[cfg(test)]
-#[path = "tests/rel_time.rs"]
-mod rel_time_tests;
-
-/// `aizen time list` — a static, glanceable print of the checkpoint timeline (newest first), with the
-/// active point marked `▸`, relative times, labels, and `auto`/`+chat` tags. Read-only, and CLI-only:
-/// in the REPL `/timemachine` shows the same history as a picker, so there is nothing to print.
-fn print_timeline() -> Result<()> {
-    let (snaps, cursor) = timemachine::timeline()?;
-    if snaps.is_empty() {
-        tui::emit_line(
-            &style("⎇ timeline — no checkpoints yet · /checkpoint to save one")
-                .dim()
-                .to_string(),
-        );
-        return Ok(());
-    }
-    let n = snaps.len();
-    tui::emit_line(&format!(
-        "{}  {n} checkpoint(s)",
-        style("⎇ timeline").color256(splash::ACCENT).bold(),
-    ));
-    // Align the `#id` column to the widest id present (+1 for the leading `#`).
-    let id_w = snaps
-        .iter()
-        .map(|s| s.id.to_string().len())
-        .max()
-        .unwrap_or(1)
-        + 1;
-    // Newest first (the ledger stores oldest → newest).
-    for (i, s) in snaps.iter().enumerate().rev() {
-        let is_cur = Some(i) == cursor;
-        let id = format!("#{}", s.id);
-        let rel = rel_time(&s.created);
-        let label = if s.label.is_empty() {
-            "(no label)".to_string()
-        } else {
-            s.label.clone()
-        };
-        let mut tags = String::new();
-        if s.auto {
-            tags.push_str(" · auto");
-        }
-        if s.has_chat {
-            tags.push_str(" · +chat");
-        }
-        let head = format!("{id:<id_w$}  {rel:<9}  {label}");
-        let mark = if is_cur { "▸" } else { " " };
-        // Current point accented; tags always dim. Style the marker+head as one segment, then append
-        // the dim tags separately so no ANSI code nests inside another.
-        let body = if is_cur {
-            format!(
-                "{} {}",
-                style(mark).color256(splash::ACCENT).bold(),
-                style(head).color256(splash::ACCENT)
-            )
-        } else {
-            format!("{mark} {head}")
-        };
-        let tag_str = if tags.is_empty() {
-            String::new()
-        } else {
-            style(tags).dim().to_string()
-        };
-        tui::emit_line(&format!("{body}{tag_str}"));
-    }
-    tui::emit_line(
-        &style("▸ = current · restore: aizen time restore <id>   (or /timemachine in the REPL)")
-            .dim()
-            .to_string(),
-    );
-    Ok(())
-}
-
-/// `/timemachine` — the whole time machine in one list: every checkpoint, and picking one rewinds to
-/// that state.
-///
-/// A row carries the id, how long ago it was taken, its label, and the `+chat` tag when the
-/// conversation was captured alongside the tree. Picking a row restores everything that checkpoint
-/// holds — the working tree always, and the conversation too whenever a chat sidecar exists — so one
-/// pick returns you to that code AND that chat. There is deliberately no Files/Task/Both sub-menu and
-/// no `pick`/`restore` argument: this list IS the surface.
-///
-/// Every restore is reversible: the pre-restore tree is auto-snapshotted, and the live conversation is
-/// saved to its own session file before being replaced.
-pub(crate) async fn timemachine_menu(
-    history: &mut Vec<Message>,
-    model_label: &mut String,
-) -> Result<()> {
-    let theme = ui_theme();
-    loop {
-        let (snaps, cursor) = match timemachine::timeline() {
-            Ok(t) => t,
-            Err(e) => {
-                println!("{e}");
-                return Ok(());
-            }
-        };
-        let n = snaps.len();
-        let mut items: Vec<String> = snaps
-            .iter()
-            .enumerate()
-            .map(|(i, s)| {
-                let here = if Some(i) == cursor { "▸ " } else { "  " };
-                let label = if s.label.is_empty() {
-                    "(no label)".to_string()
-                } else {
-                    s.label.clone()
-                };
-                // Spell out what a pick will rewind, since the pick itself is now the whole gesture.
-                let scope = if s.has_chat {
-                    "code + chat"
-                } else {
-                    "code only"
-                };
-                let tags = if s.auto {
-                    format!(" · auto · {scope}")
-                } else {
-                    format!(" · {scope}")
-                };
-                format!(
-                    "{here}#{} {}  {label}{}",
-                    s.id,
-                    style(rel_time(&s.created)).dim(),
-                    style(tags).dim(),
-                )
-            })
-            .collect();
-        items.push("✚ Save a checkpoint now (code + chat)".to_string());
-        items.push("Back".to_string());
-        let prompt = format!(
-            "Time machine — {n} checkpoint(s); pick one to rewind to it (reversible). Esc to go back"
-        );
-        let pick = match Select::with_theme(&theme)
-            .with_prompt(prompt)
-            .items(&items)
-            .default(cursor.unwrap_or(0))
-            .interact_opt()?
-        {
-            Some(i) => i,
-            None => return Ok(()),
-        };
-        if pick < n {
-            restore_checkpoint(&snaps[pick], history, model_label)?;
-        } else if pick == n {
-            let label: String = Input::with_theme(&theme)
-                .with_prompt("Label (optional)")
-                .allow_empty(true)
-                .interact_text()?;
-            // Capture the conversation alongside the tree so this checkpoint supports task restore.
-            match timemachine::save_with_chat(label.trim(), false, history) {
-                Ok(s) => println!(
-                    "{} #{} ({})",
-                    style("✓ checkpoint").color256(splash::ACCENT),
-                    s.id,
-                    if s.has_chat {
-                        "code + chat"
-                    } else {
-                        "files only"
-                    }
-                ),
-                Err(e) => println!("{}", style(format!("save failed: {e}")).red()),
-            }
-        } else {
-            return Ok(());
-        }
-    }
-}
-
-/// Rewind to everything a checkpoint holds: the working tree, plus the conversation when that
-/// checkpoint captured one.
-///
-/// There is no Files / Task / Both question any more — picking a row in the time machine means "put me
-/// back there", and what that restores is a property of the checkpoint, not a choice to re-litigate.
-/// A `/checkpoint` (or one saved from the picker) carries a chat sidecar and rewinds code + chat; an
-/// auto/agent checkpoint has no sidecar and rewinds code only, which the row already says.
-fn restore_checkpoint(
-    snap: &timemachine::Snapshot,
-    history: &mut Vec<Message>,
-    model_label: &mut String,
-) -> Result<()> {
-    // Files-only checkpoints (auto/agent) have no chat to restore.
-    if !snap.has_chat {
-        return files_restore(snap.id);
-    }
-    // Preflight and durably back up chat BEFORE files move. The live `history` is only assigned
-    // after file restore succeeds, so a failed files phase cannot leave files/chat divergent.
-    let chat = timemachine::load_chat_checked(snap.id)?;
-    if chat.is_empty() {
-        bail!("checkpoint #{} has an empty saved conversation", snap.id);
-    }
-    let backup = current_session_slug().unwrap_or_else(|| allocate_session_slug(history));
-    save_session(history, &backup, Some(model_label))
-        .context("backing up the current conversation before restore")?;
-    files_restore(snap.id)?;
-    *history = chat;
-    migrate_legacy_prompt_lanes(history, model_label);
-    refresh_prompt_lanes_for_thread_switch(history, model_label);
-    // The rewound thread continues under a NEW file — keeping the old slug would make the
-    // next autosave overwrite the backup that was just written.
-    set_session_slug(None);
-    update_live_history(history);
-    println!(
-        "{} #{} — files and conversation rewound",
-        style("⏪ restored").color256(splash::ACCENT),
-        snap.id
-    );
-    println!(
-        "{}",
-        style(format!(
-            "  (your previous chat was saved as “{backup}” — /sessions to get it back)"
-        ))
-        .dim()
-    );
-    Ok(())
-}
-
-/// Rewind only the working tree to checkpoint `id` (reversible — pre-restore tree auto-saved).
-fn files_restore(id: u32) -> Result<()> {
-    let s = timemachine::restore(id).with_context(|| format!("restoring checkpoint #{id}"))?;
-    println!(
-        "{} #{} — files rewound; your chat is untouched",
-        style("⏪ restored").color256(splash::ACCENT),
-        s.id
-    );
-    println!(
-        "{}",
-        style("  (reversible — the pre-restore tree was auto-saved; pick it to go back)").dim()
-    );
-    Ok(())
-}
-
 // ───────────────────────────── interactive landing menu ─────────────────────────────
 // Bare `ng` (no subcommand) drops into a colored, arrow-key TUI (dialoguer): a status banner +
 // a Select list (Setup / Chat / Agent / Models / Memory / Quit). ↑/↓ + Enter to choose, Esc to
@@ -1966,253 +689,6 @@ pub(crate) fn ui_theme() -> ColorfulTheme {
             .bold(),
         inactive_item_prefix: style(" ".to_string()).for_stderr(),
         ..ColorfulTheme::default()
-    }
-}
-
-/// A bordered single-line input box (the "chat box") read key-by-key via `console` (raw mode), so
-/// the box redraws as you type and the cursor sits inside it. A small line editor:
-/// - type / **Backspace** / **Del** insert+delete at the cursor; **←/→** move; **Home/End** jump;
-/// - **↑/↓** walk `history` (most-recent first; ↓ past the newest restores your in-progress draft);
-/// - **Enter** submits; **Esc** clears the line AND any attached images (quits only when both are
-///   already empty); **Ctrl-C/Ctrl-D** quit.
-/// - **Attach an image** (vision) two ways (Ctrl-V can't be used — Windows Terminal eats it):
-///   **Ctrl-O** grabs a copied screenshot from the clipboard (Win+Shift+S), or **drag an image file
-///   onto the window** (the terminal pastes its path; the caller turns image-file paths on the line
-///   into attachments on Enter). An `[N img]` tag shows in the top border; **Ctrl-X** removes the
-///   most recent attachment (keeps your text).
-///
-/// Returns `Some((line, images))` on Enter (`images` = `data:` URLs of clipboard attachments; the
-/// caller adds any file-path attachments), or `None` to quit (Esc-empty / Ctrl-C/D / EOF / non-TTY).
-/// The visible window scrolls horizontally so the cursor stays in view on long lines.
-fn read_input_box(history: &[String]) -> Result<Option<(String, Vec<String>)>> {
-    use console::{Key, Term};
-    use std::io::Write;
-    const W: usize = 66; // inner width between the │ borders
-    let text_cols = W - 3; // columns for editable text (after " ❯ ")
-
-    let term = Term::stdout();
-    let accent = splash::ACCENT;
-    let bar = |l: &str, r: &str| {
-        style(format!("{l}{}{r}", "─".repeat(W)))
-            .color256(accent)
-            .to_string()
-    };
-    // A small status tag in the TOP border (`╭───────[1 img]─╮`). ASCII-only + right-aligned, so the
-    // width is exact and the border never tears (an emoji caption mis-measures by a cell). Empty tag
-    // → a plain border.
-    let top_bar = |tag: &str| -> String {
-        if tag.is_empty() {
-            return bar("╭", "╮");
-        }
-        let t = format!("[{tag}]");
-        let fill = W.saturating_sub(t.chars().count() + 1);
-        style(format!("╭{}{t}─╮", "─".repeat(fill)))
-            .color256(accent)
-            .to_string()
-    };
-    // Attachment count → tag text (empty when none, so the border goes plain).
-    let count_tag = |n: usize| -> String {
-        if n == 0 {
-            String::new()
-        } else {
-            format!("{n} img")
-        }
-    };
-
-    // Render the middle line for (chars, cursor), scrolling so the cursor is visible. Returns the
-    // line + how far left to shift the cursor from the line end to land on `cursor`. (Char widths
-    // are treated as 1 — fine for ASCII/Latin/Vietnamese; exotic wide input may wobble by a cell.)
-    let render = |chars: &[char], cursor: usize, scroll: &mut usize| -> (String, usize) {
-        if cursor < *scroll {
-            *scroll = cursor;
-        }
-        if cursor >= *scroll + text_cols {
-            *scroll = cursor + 1 - text_cols;
-        }
-        let end = (*scroll + text_cols).min(chars.len());
-        let shown: String = chars[*scroll..end].iter().collect();
-        let shown_w = end - *scroll;
-        let pad = text_cols - shown_w;
-        let line = format!(
-            "{l} {arrow} {shown}{sp}{l}",
-            l = style("│").color256(accent),
-            arrow = style("❯").color256(accent).bold(),
-            sp = " ".repeat(pad)
-        );
-        let cursor_col = cursor - *scroll;
-        let back = (shown_w - cursor_col) + pad + 1; // chars after cursor + pad + right border
-        (line, back)
-    };
-
-    let mut scroll = 0usize;
-    let (mid0, back0) = render(&[], 0, &mut scroll);
-    println!("{}", top_bar(""));
-    println!("{mid0}");
-    print!("{}", bar("╰", "╯"));
-    std::io::stdout().flush().ok();
-    term.move_cursor_up(1).ok();
-
-    let place = |line: &str, back: usize| {
-        let _ = term.clear_line();
-        let mut o = std::io::stdout();
-        let _ = write!(o, "\r{line}");
-        let _ = o.flush();
-        let _ = term.move_cursor_left(back);
-    };
-    place(&mid0, back0);
-
-    // Repaint the TOP border (cursor sits on the middle line) — used by the image attach/remove keys
-    // to reflect the count tag, then return to the middle line (the loop's `place` restores the
-    // cursor column).
-    let redraw_top = |s: &str| {
-        let _ = term.move_cursor_up(1);
-        let _ = term.clear_line();
-        let mut o = std::io::stdout();
-        let _ = write!(o, "\r{s}");
-        let _ = o.flush();
-        let _ = term.move_cursor_down(1);
-    };
-
-    let mut chars: Vec<char> = Vec::new();
-    let mut cursor = 0usize;
-    let mut hist_idx: Option<usize> = None; // Some = currently browsing history
-    let mut draft: Vec<char> = Vec::new(); // the in-progress line saved when entering history
-    let mut images: Vec<String> = Vec::new(); // pasted vision attachments (data: URLs)
-
-    loop {
-        let key = match term.read_key() {
-            Ok(k) => k,
-            Err(_) => return Ok(None),
-        };
-        match key {
-            Key::Enter => {
-                let text: String = chars.iter().collect();
-                // Collapse the 3-line box into a single compact `> …` echo (nothing when empty), so
-                // the scrollback reads as a clean transcript instead of a stack of empty boxes — AND
-                // so the box's presence is the unambiguous "your turn to type" signal (no box +
-                // spinner/⊙ traces = the agent is working).
-                term.move_cursor_down(1).ok(); // → bottom border
-                term.clear_line().ok();
-                term.move_cursor_up(1).ok(); // → middle line
-                term.clear_line().ok();
-                term.move_cursor_up(1).ok(); // → top border
-                term.clear_line().ok();
-                print!("\r");
-                let trimmed = text.trim();
-                if !trimmed.is_empty() {
-                    println!(
-                        "{} {}",
-                        style("❯").color256(accent).bold(),
-                        style(trimmed).dim()
-                    );
-                } else if !images.is_empty() {
-                    println!(
-                        "{} {}",
-                        style("❯").color256(accent).bold(),
-                        style(format!("📎 {} image(s)", images.len())).dim()
-                    );
-                }
-                std::io::stdout().flush().ok();
-                return Ok(Some((text, images)));
-            }
-            Key::Char('\u{f}') => {
-                // Ctrl-O: grab a copied screenshot from the clipboard (Win+Shift+S / "Copy image").
-                // Explicit, so it works in Windows Terminal (which eats Ctrl-V but forwards Ctrl-O).
-                let tag = match image_input::clipboard_image_data_url() {
-                    Ok(Some(url)) => {
-                        images.push(url);
-                        count_tag(images.len())
-                    }
-                    Ok(None) => "no image".to_string(),
-                    Err(_) => "clip error".to_string(),
-                };
-                redraw_top(&top_bar(&tag));
-            }
-            Key::Char('\u{18}') => {
-                // Ctrl-X: remove the most recently attached image (keeps your typed text). The tag
-                // reflects the new count (gone when the last one is removed); no-op when none.
-                if images.pop().is_some() {
-                    redraw_top(&top_bar(&count_tag(images.len())));
-                }
-            }
-            Key::Escape => {
-                // Nothing typed AND nothing attached → quit. Otherwise clear the line AND drop any
-                // attached images (a quick way to start over / undo a wrong attachment).
-                if chars.is_empty() && images.is_empty() {
-                    term.move_cursor_down(1).ok();
-                    println!();
-                    return Ok(None);
-                }
-                chars.clear();
-                cursor = 0;
-                hist_idx = None;
-                if !images.is_empty() {
-                    images.clear();
-                    redraw_top(&top_bar(""));
-                }
-            }
-            Key::Char('\u{3}') | Key::Char('\u{4}') => {
-                term.move_cursor_down(1).ok();
-                println!();
-                return Ok(None);
-            }
-            Key::Char(c) if c.is_control() => continue, // ignore stray control chars (no redraw)
-            Key::Char(c) => {
-                chars.insert(cursor, c);
-                cursor += 1;
-            }
-            Key::Backspace => {
-                if cursor > 0 {
-                    chars.remove(cursor - 1);
-                    cursor -= 1;
-                }
-            }
-            Key::Del => {
-                if cursor < chars.len() {
-                    chars.remove(cursor);
-                }
-            }
-            Key::ArrowLeft => cursor = cursor.saturating_sub(1),
-            Key::ArrowRight => {
-                if cursor < chars.len() {
-                    cursor += 1;
-                }
-            }
-            Key::Home => cursor = 0,
-            Key::End => cursor = chars.len(),
-            Key::ArrowUp => {
-                if history.is_empty() {
-                    continue;
-                }
-                let next = match hist_idx {
-                    None => {
-                        draft = chars.clone(); // save the in-progress line
-                        history.len() - 1
-                    }
-                    Some(0) => continue, // already at the oldest
-                    Some(i) => i - 1,
-                };
-                hist_idx = Some(next);
-                chars = history[next].chars().collect();
-                cursor = chars.len();
-            }
-            Key::ArrowDown => match hist_idx {
-                None => continue,
-                Some(i) if i + 1 < history.len() => {
-                    hist_idx = Some(i + 1);
-                    chars = history[i + 1].chars().collect();
-                    cursor = chars.len();
-                }
-                Some(_) => {
-                    hist_idx = None; // past the newest → restore the draft
-                    chars = draft.clone();
-                    cursor = chars.len();
-                }
-            },
-            _ => continue, // unhandled key → no redraw
-        }
-        let (m, b) = render(&chars, cursor, &mut scroll);
-        place(&m, b);
     }
 }
 
@@ -3905,134 +2381,6 @@ fn memory_auto_learn_enabled() -> bool {
     cli_config::load().memory_auto_learn.unwrap_or(true)
 }
 
-/// Decide THIS turn's reasoning effort. When auto-detect is ON (default), classify the user's
-/// fully-expanded message (keyword ladder → complexity heuristic); a hit forces that tier, else we
-/// fall back to the configured `reasoning_effort` (which may itself be `None` ⇒ omit the field).
-/// PURE-ish wrapper (loads config) around the pure `core::effort::classify_effort`. The result is
-/// armed into the per-turn override cell read by the LLM client, and cleared at turn end.
-fn resolve_turn_effort(line: &str) -> Option<String> {
-    // Ultimate mode pins max effort every turn (auto-detect is bypassed) — the aizen `ultracode`.
-    if cli_config::ultimate_enabled() {
-        return Some("max".to_string());
-    }
-    if cli_config::auto_effort_enabled() {
-        // P3: adaptive routing lets the complexity heuristic climb to `xhigh` on the hardest turns.
-        let adaptive = cli_config::adaptive_effort_enabled();
-        if let Some(e) = crate::core::effort::classify_effort_with(line, adaptive) {
-            return Some(e.as_str().to_string());
-        }
-    }
-    cli_config::load().reasoning_effort.clone()
-}
-
-/// The per-turn "effort: <tier>" status line, tinted to match the slider's tier colours (auto =
-/// moonlight · low = green · medium = dim silver · high = gold) so the whole effort feature reads as
-/// one system. `None` ⇒ the field is omitted this turn → shown as a faint "default".
-fn effort_turn_line(eff: Option<&str>) -> String {
-    // low = green, medium = dim silver; the three "hot" rungs escalate high → xhigh → max
-    // (gold → bold gold → salmon) so the eye can tell them apart at a glance.
-    let styled = match eff {
-        Some("low") => console::style("low".to_string()).color256(theme::OK),
-        Some("medium") => console::style("medium".to_string()).color256(theme::ACCENT_DIM),
-        Some("high") => console::style("high".to_string()).color256(theme::WARN),
-        Some("xhigh") => console::style("xhigh".to_string())
-            .color256(theme::WARN)
-            .bold(),
-        Some("max") => console::style("max".to_string())
-            .color256(theme::ERR)
-            .bold(),
-        Some(other) => console::style(other.to_string()).color256(theme::ACCENT),
-        None => console::style("default".to_string()).color256(theme::FAINT),
-    };
-    format!("{} {}", theme::faint("  effort:"), styled)
-}
-
-/// The current effort setting as a slider index: 0 = auto (auto-detect ON, no pinned tier), else the
-/// pinned tier (1=low · 2=medium · 3=high). A pinned-but-unknown effort string, or auto-off with no
-/// pin, both fall back to `auto` so the slider always opens on a valid stop.
-fn effort_slider_start() -> usize {
-    let cfg = cli_config::load();
-    if cli_config::auto_effort_enabled() {
-        return 0; // auto ON ⇒ the "auto" stop, regardless of any stale pinned value
-    }
-    match cfg.reasoning_effort.as_deref() {
-        Some("low") => 1,
-        Some("medium") => 2,
-        Some("high") => 3,
-        Some("xhigh") => 4,
-        Some("max") => 5,
-        _ => 0,
-    }
-}
-
-/// Apply a slider choice to the config and persist it. `0` ⇒ auto (auto_effort=None, clear the pin);
-/// `1..=5` ⇒ pin low/medium/high/xhigh/max and turn auto off — the exact same writes as `/effort auto`
-/// and `/effort low|medium|high|xhigh|max`, so the slider and the text commands stay in lockstep.
-fn apply_effort_choice(idx: usize) {
-    let mut cfg = cli_config::load();
-    let msg = match idx {
-        1..=5 => {
-            let tier = ["", "low", "medium", "high", "xhigh", "max"][idx];
-            cfg.reasoning_effort = Some(tier.to_string());
-            cfg.auto_effort = Some(false);
-            format!("effort pinned to {tier} (auto off) — every turn now sends reasoning_effort={tier}.")
-        }
-        _ => {
-            cfg.auto_effort = None; // None ⇒ auto ON (the default)
-            cfg.reasoning_effort = None; // clear any stale pin so auto isn't shadowed
-            "effort auto ON — each turn's effort is detected from your message (keyword + complexity).".to_string()
-        }
-    };
-    match cli_config::save(&cfg) {
-        Ok(_) => tui::emit_line(&style(msg).color256(splash::ACCENT).to_string()),
-        Err(e) => tui::emit_line(&format!("{} {e}", style("effort:").red())),
-    }
-}
-
-/// The plain text status report for `/effort status` (and the off-TTY fallback of the bare `/effort`).
-pub(crate) fn effort_status_report() {
-    let cfg = cli_config::load();
-    let auto = if cli_config::auto_effort_enabled() {
-        "on"
-    } else {
-        "off"
-    };
-    let fixed = cfg
-        .reasoning_effort
-        .as_deref()
-        .unwrap_or("(none — omitted)");
-    tui::emit_line(
-        &style(format!(
-            "effort: auto-detect {auto} · fixed reasoning_effort {fixed}\n\
-             /effort auto|off · /effort low|medium|high (pins it, turns auto off) · /effort none (clear)"
-        ))
-        .dim()
-        .to_string(),
-    );
-    if std::env::var("AIZEN_AUTO_EFFORT").is_ok() {
-        tui::emit_line(
-            &style("(note: AIZEN_AUTO_EFFORT is set — it overrides the auto toggle)")
-                .dim()
-                .to_string(),
-        );
-    }
-}
-
-/// Bare `/effort` → the animated drag slider. Opens on the current setting; a commit persists the
-/// choice, Esc keeps things as-is. Off-TTY the slider returns `None` immediately, so we fall back to
-/// the text report instead of leaving the user with no output.
-pub(crate) fn effort_slider_flow() {
-    use std::io::IsTerminal;
-    if !std::io::stdout().is_terminal() {
-        effort_status_report();
-        return;
-    }
-    match tui::effort_slider(effort_slider_start()) {
-        Some(idx) => apply_effort_choice(idx),
-        None => tui::emit_line(&style("(effort unchanged)").dim().to_string()),
-    }
-}
-
 /// After a completed turn: passively learn durable user/project facts from the user's last message.
 /// FREE — regex extraction, no model call — through the SAME pipeline as `aizen memory learn`
 /// (sanitize-to-fact → write-time threat-scan → confidence-route → consolidate → store, with
@@ -4619,362 +2967,6 @@ pub(crate) async fn handoff_now(history: &[Message], goal: &str) -> Result<Strin
         anyhow::bail!("the model returned an empty handoff summary");
     }
     Ok(summary.trim().to_string())
-}
-
-/// `/memory <sub>` — the in-REPL view of the same store the agent writes through, so the user can
-/// audit and correct it without dropping to the CLI. Sub-commands mirror `aizen memory <sub>` 1:1
-/// (same functions, same ids) rather than reimplementing a second, drifting surface.
-///
-/// `forget` here is the SOFT delete (archive → restorable); hard `purge` is CLI-only on purpose, so
-/// nothing typed mid-chat can destroy a fact irreversibly.
-pub(crate) async fn sessions_menu(history: &mut Vec<Message>, model_label: &str) -> Result<()> {
-    loop {
-        let theme = ui_theme();
-        // Newest first, with provenance: age, origin project for foreign/unlabeled files, a
-        // "● current" marker on the live conversation's file, and corrupt files called out as
-        // unreadable instead of posing as plausible "0 msgs" sessions.
-        let pool = scan_sessions();
-        let names: Vec<String> = pool.iter().map(|s| s.name.clone()).collect();
-        let n_sessions = pool.len();
-        let current = current_session_slug();
-        let mut items: Vec<String> = pool
-            .iter()
-            .map(|s| {
-                let count = match s.msgs {
-                    Some(n) => format!("{n} msg{}", if n == 1 { "" } else { "s" }),
-                    None => "(unreadable)".to_string(),
-                };
-                let mut row = format!(
-                    "{} {}  —  {count} · {}",
-                    icons::g(icons::slash("sessions")),
-                    s.name,
-                    fmt_session_age(s.mtime_ms)
-                );
-                if s.here == Some(false) {
-                    row.push_str(&format!(" · {}", session_origin_label(s.meta.as_ref())));
-                }
-                if current.as_deref() == Some(s.name.as_str()) {
-                    row.push_str(" · ● current");
-                }
-                row
-            })
-            .collect();
-        items.push("+ Save current conversation…".to_string());
-        if n_sessions > 0 {
-            items.push("Delete a session".to_string());
-        }
-        items.push("Back".to_string());
-
-        let prompt = if n_sessions == 0 {
-            "Sessions — none saved yet (Esc to go back)".to_string()
-        } else {
-            format!("Sessions — {n_sessions} saved, newest first · pick one to restore (Esc to go back)")
-        };
-        let pick = match Select::with_theme(&theme)
-            .with_prompt(prompt)
-            .items(&items)
-            .default(0)
-            .interact_opt()?
-        {
-            Some(i) => i,
-            None => return Ok(()),
-        };
-
-        if pick < n_sessions {
-            let name = &names[pick];
-            match load_session(history, name, model_label) {
-                Ok(n) => {
-                    reset_per_session_state(); // thread switch — same contract as /resume
-                    println!(
-                        "{}",
-                        style(format!(
-                            "restored '{}' ({n} messages)",
-                            pretty_session_name(name)
-                        ))
-                        .color256(splash::ACCENT)
-                    );
-                    agent::replay_transcript(history);
-                    return Ok(());
-                }
-                Err(e) => tui::note_line(&format!("{} {e}", style("restore:").red())),
-            }
-        } else if items[pick].starts_with("+ Save") {
-            let suggested = suggest_session_name(history);
-            let name: String = Input::with_theme(&theme)
-                .with_prompt("Save as")
-                .with_initial_text(suggested)
-                .interact_text()
-                .unwrap_or_default();
-            if !name.trim().is_empty() {
-                // Saving over a DIFFERENT conversation's file is destructive — confirm it. Re-saving
-                // the live conversation's own file is the normal case and stays silent.
-                let target = sanitize_name(name.trim());
-                if let Some(why) = session_save_name_error(&name) {
-                    eprintln!("{} {why}", style("save:").red());
-                    continue;
-                }
-                let exists = sessions_dir().join(format!("{target}.json")).exists();
-                if exists && current.as_deref() != Some(target.as_str()) {
-                    let overwrite = Confirm::with_theme(&theme)
-                        .with_prompt(format!(
-                            "'{}' already exists — overwrite it?",
-                            pretty_session_name(&target)
-                        ))
-                        .default(false)
-                        .interact_opt()?
-                        .unwrap_or(false);
-                    if !overwrite {
-                        continue;
-                    }
-                }
-                match save_session(history, name.trim(), Some(model_label)) {
-                    Ok(_) => {
-                        // Pin the session to this name so later autosaves keep rewriting the SAME file
-                        // (both paths route through `sanitize_name`, so the raw name maps to one file).
-                        set_session_slug(Some(name.trim().to_string()));
-                        update_live_history(history); // exit-flush snapshot follows the pinned name
-                        println!(
-                            "{}",
-                            style(format!("saved '{}'", name.trim())).color256(splash::ACCENT)
-                        );
-                    }
-                    Err(e) => tui::note_line(&format!("{} {e}", style("save:").red())),
-                }
-            }
-        } else if items[pick] == "Delete a session" {
-            if let Ok(Some(i)) = Select::with_theme(&theme)
-                .with_prompt("Delete which session? (Esc to cancel)")
-                .items(&names)
-                .default(0)
-                .interact_opt()
-            {
-                let slug = &names[i];
-                let pretty = pretty_session_name(slug);
-                let confirmed = Confirm::with_theme(&theme)
-                    .with_prompt(format!("Delete '{pretty}' permanently?"))
-                    .default(false)
-                    .interact_opt()?
-                    .unwrap_or(false);
-                if !confirmed {
-                    continue;
-                }
-                match delete_session(slug) {
-                    Ok(_) => {
-                        if current_session_slug().as_deref() == Some(slug.as_str()) {
-                            set_session_slug(None);
-                        }
-                        println!(
-                            "{}",
-                            style(format!("deleted '{pretty}'")).color256(splash::ACCENT)
-                        );
-                    }
-                    Err(e) => tui::note_line(&format!("{} {e}", style("delete:").red())),
-                }
-            }
-        } else {
-            return Ok(()); // Back
-        }
-    }
-}
-
-/// `/import` — pick a conversation recorded by Claude Code or Codex (for THIS project) and resume
-/// it inside aizen. The foreign transcript is parsed into `Vec<Message>`, repaired so it satisfies
-/// `assert_valid_history`, then handed to the SAME thread-switch path `/resume` uses: refresh the
-/// prompt lanes for the current project + model, reset per-session state, and replay so the
-/// restored thread is VISIBLE rather than silently present.
-///
-/// Foreign transcripts are never autosaved back over themselves — the imported conversation becomes
-/// the live one and is autosaved under a fresh aizen slug from then on, exactly like a `/resume`
-/// of an aizen session. The source file is only ever READ.
-pub(crate) async fn import_menu(history: &mut Vec<Message>, model_label: &str) -> Result<()> {
-    let theme = ui_theme();
-    let pool = features::foreign_session::discover(&config::project_root());
-    if pool.is_empty() {
-        tui::emit_line(
-            &style("no Claude Code or Codex transcripts found for this project")
-                .dim()
-                .to_string(),
-        );
-        tui::emit_line(
-            &style("(they appear here once you've used `claude` or `codex` in this directory)")
-                .dim()
-                .to_string(),
-        );
-        return Ok(());
-    }
-    // Clip to the terminal, minus dialoguer's own `❯ ` prefix and one cell of right margin. An item
-    // that overflows gets WRAPPED onto a second line, which breaks the column alignment for every row
-    // below it — the whole reason the list is scannable.
-    let width = console::Term::stdout().size().1 as usize;
-    let items: Vec<String> = pool
-        .iter()
-        .map(|s| s.row(fmt_session_age_compact, width.saturating_sub(4).max(24)))
-        .collect();
-    let prompt = format!("Import — {} conversations from claude/codex", pool.len());
-    let pick = match Select::with_theme(&theme)
-        .with_prompt(prompt)
-        .items(&items)
-        .default(0)
-        .interact_opt()?
-    {
-        Some(i) => i,
-        None => return Ok(()),
-    };
-    let session = &pool[pick];
-    match features::foreign_session::load(session) {
-        Ok(imported) => {
-            // Drop any system lanes the source CLI's harness left in (already filtered in parse,
-            // but defend against a future schema that embeds them) before splicing aizen's own.
-            *history = imported;
-            refresh_prompt_lanes_for_thread_switch(history, model_label);
-            // Same thread-switch contract as /resume: the foreign thread's todos/cost/grants
-            // belong to it, not to whatever was live before the import.
-            reset_per_session_state();
-            set_session_slug(None); // the imported chat autosaves under a fresh aizen slug from here
-            agent::replay_transcript(history);
-            let tag = match session.cli {
-                features::foreign_session::Cli::Claude => "claude",
-                features::foreign_session::Cli::Codex => "codex",
-            };
-            tui::emit_line(
-                &style(format!(
-                    "⇲ imported “{}” from {} — {} messages, context restored",
-                    if session.first_prompt.is_empty() {
-                        "(no prompt)"
-                    } else {
-                        &session.first_prompt
-                    },
-                    tag,
-                    history.len()
-                ))
-                .color256(splash::ACCENT)
-                .to_string(),
-            );
-            tui::emit_line(
-                &style(format!("  source: {}", session.path.display()))
-                    .dim()
-                    .to_string(),
-            );
-        }
-        Err(e) => tui::emit_line(&format!("{} {e}", style("import:").red())),
-    }
-    Ok(())
-}
-
-/// `aizen import [path]` — CLI surface. No path lists every foreign transcript for this project;
-/// a path loads that file's transcript and prints a one-line summary (the CLI can't resume into a
-/// REPL, so it reports what WOULD be loaded — the actual resume happens via `/import` in the REPL).
-async fn run_import(path: Option<String>) -> Result<()> {
-    match path {
-        Some(p) => {
-            let p = std::path::PathBuf::from(&p);
-            let bytes = std::fs::read(&p).with_context(|| format!("reading {}", p.display()))?;
-            // Detect the CLI from the file's own shape rather than the path: a Claude line has a
-            // top-level `type` that is "user"/"assistant"/"mode"/…; a Codex line has `session_meta`
-            // or `response_item`. Sniff the first parseable line.
-            let cli = sniff_cli(&bytes);
-            let sess = features::foreign_session::ForeignSession {
-                cli,
-                path: p.clone(),
-                cwd: String::new(),
-                mtime_ms: None,
-                turns: 0,
-                first_prompt: String::new(),
-            };
-            match features::foreign_session::load(&sess) {
-                Ok(msgs) => {
-                    let tag = match cli {
-                        features::foreign_session::Cli::Claude => "claude",
-                        features::foreign_session::Cli::Codex => "codex",
-                    };
-                    println!(
-                        "{}",
-                        style(format!(
-                            "⇲ {} transcript — {} messages ready to resume",
-                            tag,
-                            msgs.len()
-                        ))
-                        .color256(splash::ACCENT)
-                    );
-                    println!("  source: {}", p.display());
-                    println!(
-                        "  {}",
-                        style("open the REPL in this project and run /import to resume it").dim()
-                    );
-                    Ok(())
-                }
-                Err(e) => anyhow::bail!("import: {e}"),
-            }
-        }
-        None => {
-            let pool = features::foreign_session::discover(&config::project_root());
-            if pool.is_empty() {
-                println!(
-                    "{}",
-                    style("no Claude Code or Codex transcripts found for this project").dim()
-                );
-                println!(
-                    "{}",
-                    style(
-                        "(they appear here once you've used `claude` or `codex` in this directory)"
-                    )
-                    .dim()
-                );
-                return Ok(());
-            }
-            println!(
-                "{}",
-                style(format!(
-                    "Foreign transcripts for this project ({}), newest first:",
-                    pool.len()
-                ))
-                .color256(splash::ACCENT)
-            );
-            for s in &pool {
-                let tag = match s.cli {
-                    features::foreign_session::Cli::Claude => "claude",
-                    features::foreign_session::Cli::Codex => "codex",
-                };
-                println!(
-                    "  [{}] {:>6} · {} turns · {}",
-                    tag,
-                    fmt_session_age(s.mtime_ms),
-                    s.turns,
-                    s.path.display()
-                );
-            }
-            println!(
-                "{}",
-                style("resume one with: /import  (in the REPL)  or  aizen import <path>").dim()
-            );
-            Ok(())
-        }
-    }
-}
-
-/// Decide which CLI a transcript belongs to from its content, not its path. Falls back to Claude
-/// (the more permissive parser — it ignores unrecognized line types) when the sniff is inconclusive.
-fn sniff_cli(bytes: &[u8]) -> features::foreign_session::Cli {
-    for line in bytes.split(|b| *b == b'\n') {
-        if line.is_empty() {
-            continue;
-        }
-        let Ok(v) = serde_json::from_slice::<serde_json::Value>(line) else {
-            continue;
-        };
-        let ty = v.get("type").and_then(|t| t.as_str()).unwrap_or("");
-        if ty == "session_meta"
-            || ty == "response_item"
-            || ty == "event_msg"
-            || ty == "turn_context"
-        {
-            return features::foreign_session::Cli::Codex;
-        }
-        if ty == "user" || ty == "assistant" || ty == "mode" || ty == "file-history-snapshot" {
-            return features::foreign_session::Cli::Claude;
-        }
-    }
-    features::foreign_session::Cli::Claude
 }
 
 /// Switch to one saved provider profile. Root endpoint fields are updated atomically, so the next
@@ -5840,1164 +3832,6 @@ fn run_memory_model_list() -> Result<()> {
             "note: this build has NO dense feature — rebuild with `--features dense` to use it."
         );
     }
-    Ok(())
-}
-
-fn run_persona(cmd: PersonaCmd) -> Result<()> {
-    match cmd {
-        PersonaCmd::List => {
-            let active_slug = cli_config::load()
-                .persona
-                .as_deref()
-                .map(skill::sanitize_name);
-            let ps = persona::list();
-            if ps.is_empty() {
-                println!("(no personas yet — `aizen persona new <name>`, or /persona in the REPL)");
-                return Ok(());
-            }
-            for p in &ps {
-                let slug = skill::sanitize_name(&p.name);
-                let mark = if active_slug.as_deref() == Some(slug.as_str()) {
-                    "●"
-                } else {
-                    "○"
-                };
-                let sub = if p.role.is_empty() {
-                    p.voice.clone()
-                } else {
-                    p.role.clone()
-                };
-                let (eps, ins) = persona::self_mem::counts(&slug);
-                println!(
-                    "{mark} {} — {sub}  ({ins} insights, {eps} episodes)",
-                    p.name
-                );
-            }
-            // Retired cards are recoverable, so say they exist — a soft delete nobody can see is
-            // indistinguishable from a hard one.
-            let retired = persona::list_archive();
-            if !retired.is_empty() {
-                let names: Vec<&str> = retired.iter().map(|p| p.name.as_str()).collect();
-                println!(
-                    "\nretired: {} — `aizen persona restore <name>` brings one back",
-                    names.join(", ")
-                );
-            }
-            // Personas have no `where` sub-command of their own (one folder, no zoning), so the path
-            // goes here directly — the card and its `.self/` memory are plain files worth editing.
-            println!(
-                "{}",
-                style(format!(
-                    "\nfiles: {}   (`<name>.md` is the card, `<name>.self/` its memory)",
-                    persona::personas_dir().display()
-                ))
-                .dim()
-            );
-            Ok(())
-        }
-        PersonaCmd::Show { name } => {
-            let p = persona::load(&name).ok_or_else(|| anyhow::anyhow!("no persona '{name}'"))?;
-            println!("# {}", p.name);
-            if !p.role.is_empty() {
-                println!("role: {}", p.role);
-            }
-            if !p.voice.is_empty() {
-                println!("voice: {}", p.voice);
-            }
-            if !p.body.is_empty() {
-                println!("\n{}", p.body);
-            }
-            Ok(())
-        }
-        PersonaCmd::New {
-            name,
-            role,
-            voice,
-            body,
-        } => {
-            let body = match body {
-                Some(b) => b,
-                None => read_stdin("reading persona body from stdin")?,
-            };
-            let path = persona::save(
-                &name,
-                role.as_deref().unwrap_or(""),
-                voice.as_deref().unwrap_or(""),
-                &body,
-            )?;
-            println!("saved persona → {}", path.display());
-            Ok(())
-        }
-        PersonaCmd::Use { name } => {
-            let p = persona::load(&name)
-                .ok_or_else(|| anyhow::anyhow!("no persona '{name}' (see `aizen persona list`)"))?;
-            let mut cfg = cli_config::load();
-            cfg.persona = Some(p.name.clone());
-            cli_config::save(&cfg)?;
-            println!("now playing: {}", p.name);
-            Ok(())
-        }
-        PersonaCmd::Clear => {
-            let mut cfg = cli_config::load();
-            cfg.persona = None;
-            cli_config::save(&cfg)?;
-            println!("persona cleared → default assistant voice");
-            Ok(())
-        }
-        PersonaCmd::SelfMem { name, all } => {
-            let slug = match name {
-                Some(n) => skill::sanitize_name(&n),
-                None => persona::active_slug().ok_or_else(|| {
-                    anyhow::anyhow!("no active persona — pass a name or `aizen persona use <name>`")
-                })?,
-            };
-            let label = persona::load(&slug)
-                .map(|p| p.name)
-                .unwrap_or_else(|| slug.clone());
-            persona_self_view_n(&slug, &label, all);
-            Ok(())
-        }
-        PersonaCmd::Forget { id, name } => {
-            let slug = match name {
-                Some(n) => skill::sanitize_name(&n),
-                None => persona::active_slug().ok_or_else(|| {
-                    anyhow::anyhow!("no active persona — pass --name or `aizen persona use <name>`")
-                })?,
-            };
-            persona::self_mem::forget(&slug, &id)?;
-            println!(
-                "retired self-memory '{id}' (archived — `aizen persona unforget {id}` restores it)"
-            );
-            Ok(())
-        }
-        PersonaCmd::Unforget { id, name } => {
-            let slug = match name {
-                Some(n) => skill::sanitize_name(&n),
-                None => persona::active_slug().ok_or_else(|| {
-                    anyhow::anyhow!("no active persona — pass --name or `aizen persona use <name>`")
-                })?,
-            };
-            persona::self_mem::restore(&slug, &id)?;
-            println!("restored self-memory '{id}'");
-            Ok(())
-        }
-        PersonaCmd::Remember { text, importance } => {
-            let slug = persona::active_slug().ok_or_else(|| {
-                anyhow::anyhow!("no active persona — `aizen persona use <name>` first")
-            })?;
-            // Explicit CLI remember is always formative: force Explicit kind + floor ≥ FORMATIVE_MIN.
-            let imp = importance
-                .unwrap_or_else(|| {
-                    persona::self_mem::classify_turn(&text, 0)
-                        .map(|s| s.importance)
-                        .unwrap_or(6)
-                })
-                .max(persona::self_mem::FORMATIVE_MIN)
-                .min(10);
-            let body = if text.trim().starts_with("correction:")
-                || text.trim().starts_with("preference:")
-                || text.trim().starts_with("work:")
-                || text.trim().starts_with("bond:")
-                || text.trim().starts_with("explicit:")
-            {
-                text.clone()
-            } else {
-                persona::self_mem::format_episode_body(
-                    persona::self_mem::EventKind::Explicit,
-                    &text,
-                    0,
-                    "",
-                )
-            };
-            match persona::self_mem::record_episode(&slug, &body, imp)? {
-                Some(id) => println!("recorded episode '{id}' (importance {imp})"),
-                None => println!("(skipped — near-duplicate of a recent episode/insight)"),
-            }
-            Ok(())
-        }
-        PersonaCmd::Delete { name } => {
-            if persona::delete(&name)? {
-                // Clear the active pointer too, or the config keeps naming a card that is gone.
-                let mut cfg = cli_config::load();
-                if cfg
-                    .persona
-                    .as_deref()
-                    .map(|p| skill::sanitize_name(p) == skill::sanitize_name(&name))
-                    .unwrap_or(false)
-                {
-                    cfg.persona = None;
-                    cli_config::save(&cfg)?;
-                    println!("retired persona '{name}' (was active — back to the default voice)");
-                } else {
-                    println!("retired persona '{name}'");
-                }
-                println!(
-                    "card + self-memory archived under {} — `aizen persona restore {name}`",
-                    persona::archive_dir().display()
-                );
-            } else {
-                println!("no persona named '{name}'");
-            }
-            Ok(())
-        }
-        PersonaCmd::Restore { name } => {
-            let p = persona::restore(&name)?;
-            println!("restored persona '{name}' → {}", p.display());
-            Ok(())
-        }
-        PersonaCmd::Block => {
-            match persona::prompt_block() {
-                Some(p) => println!("<persona>\n{}\n</persona>", p.trim()),
-                None => {
-                    println!("(no persona active — `aizen persona use <name>`)");
-                    return Ok(());
-                }
-            }
-            match persona::self_block() {
-                Some(s) => println!("\n<self>\n{}\n</self>", s.trim()),
-                None => println!("\n(no <self> yet — the character has no self-memory; `aizen persona remember \"...\"`)"),
-            }
-            Ok(())
-        }
-    }
-}
-
-fn run_soul(cmd: Option<SoulCmd>) -> Result<()> {
-    match cmd.unwrap_or(SoulCmd::Show) {
-        SoulCmd::Show => {
-            match soul::prompt_block() {
-                Some(b) => println!("<agent_identity>\n{}\n</agent_identity>", b.trim()),
-                None if soul::exists() => println!(
-                    "(SOUL.md exists but renders nothing — it is empty or was dropped by the safety \
-                     scan; see {})",
-                    soul::soul_path().display()
-                ),
-                None => println!(
-                    "(no operating identity yet — set one with `aizen soul set` or edit {})",
-                    soul::soul_path().display()
-                ),
-            }
-            Ok(())
-        }
-        SoulCmd::Set { body } => {
-            let body = match body {
-                Some(b) => b,
-                None => read_stdin("reading SOUL body from stdin")?,
-            };
-            let path = soul::write(&body)?;
-            println!("saved operating identity → {}", path.display());
-            if soul::prompt_block().is_none() {
-                println!(
-                    "{}",
-                    style("⚠ heads up: it renders nothing — the safety scan dropped it (a credential or \
-                     injection-looking line). It will NOT be injected until fixed.")
-                        .yellow()
-                );
-            }
-            Ok(())
-        }
-        SoulCmd::Clear => {
-            if soul::clear()? {
-                println!("operating identity cleared");
-            } else {
-                println!("(no operating identity to clear)");
-            }
-            Ok(())
-        }
-        SoulCmd::Path => {
-            println!("{}", soul::soul_path().display());
-            Ok(())
-        }
-    }
-}
-
-async fn run_skill(cmd: SkillCmd) -> Result<()> {
-    match cmd {
-        SkillCmd::List { all_zones } => {
-            let skills = skill::list();
-            if skills.is_empty() && !all_zones {
-                println!(
-                    "(no skills — add one with `aizen skill add <name>`, or /skills in the REPL)"
-                );
-                return Ok(());
-            }
-            // Aligned columns, and the usage count as the triage signal: a skill the agent has never
-            // loaded is either mis-triggered (`when:` doesn't match how the task gets phrased) or not
-            // needed. Unaligned `name — description` gave no way to compare that across rows.
-            let namew = skills
-                .iter()
-                .map(|s| s.name.chars().count())
-                .max()
-                .unwrap_or(0)
-                .min(38);
-            for s in &skills {
-                let d = if s.description.is_empty() {
-                    &s.when
-                } else {
-                    &s.description
-                };
-                let tag = match s.origin {
-                    skill::SkillOrigin::Global => "",
-                    skill::SkillOrigin::Project => " [project]",
-                    skill::SkillOrigin::Repo => " [repo]",
-                };
-                let uses = if s.uses > 0 {
-                    format!("{}×", s.uses)
-                } else {
-                    "cold".to_string()
-                };
-                let pad = namew.saturating_sub(s.name.chars().count());
-                println!(
-                    "  {}{}  {:<5} {}{tag}",
-                    s.name,
-                    " ".repeat(pad),
-                    uses,
-                    elide(d, 92)
-                );
-            }
-            let cold = skills.iter().filter(|s| s.uses == 0).count();
-            println!("\n{} skill(s)", skills.len());
-            if cold > 0 {
-                println!(
-                    "{}",
-                    style(format!(
-                        "{cold} cold (never loaded — check `when:` matches how you'd phrase the task)"
-                    ))
-                    .dim()
-                );
-            }
-            println!(
-                "{}",
-                style(
-                    "`skill show <name>` reads one · `skill refine <name>` rewrites the steps · \
-                     `skill delete <name>` retires (restorable)\n\
-                     `skill where` prints the three folders skills are read from"
-                )
-                .dim()
-            );
-            if all_zones {
-                let others = skill::list_other_zones();
-                if !others.is_empty() {
-                    println!("\nother workspaces' zones (invisible here):");
-                    for (zone, s) in &others {
-                        println!("  {}  [p:{zone}]  {}", s.name, elide(&s.description, 80));
-                    }
-                }
-            }
-            let retired = skill::list_archive();
-            if !retired.is_empty() {
-                let names: Vec<&str> = retired.iter().map(|s| s.name.as_str()).collect();
-                println!(
-                    "\nretired: {} — `aizen skill restore <name>` brings one back",
-                    names.join(", ")
-                );
-            }
-            Ok(())
-        }
-        SkillCmd::Show { name } => match skill::load(&name) {
-            Some(sk) => {
-                println!("{}", skill::render_loaded(&sk));
-                Ok(())
-            }
-            None => anyhow::bail!("no skill named '{name}' (try `aizen skill list`)"),
-        },
-        SkillCmd::Where => {
-            println!("{}", skill_where_report());
-            Ok(())
-        }
-        SkillCmd::Add {
-            name,
-            description,
-            when,
-            body,
-        } => {
-            let body = match body {
-                Some(b) => b,
-                None => read_stdin("reading skill body from stdin")?,
-            };
-            let path = skill::save(
-                &name,
-                description.as_deref().unwrap_or(""),
-                when.as_deref().unwrap_or(""),
-                &body,
-            )?;
-            println!("saved skill → {}", path.display());
-            Ok(())
-        }
-        SkillCmd::Delete { name } => {
-            if skill::delete(&name)? {
-                println!(
-                    "retired '{name}' (archived — `aizen skill restore {name}` brings it back)"
-                );
-            } else {
-                println!("(no skill named '{name}')");
-            }
-            Ok(())
-        }
-        SkillCmd::Restore { name } => {
-            let p = skill::restore(&name)?;
-            println!("restored '{name}' → {}", p.display());
-            Ok(())
-        }
-        SkillCmd::Refine {
-            name,
-            description,
-            when,
-            body,
-        } => {
-            let body = match body {
-                Some(b) => b,
-                None => read_stdin("reading the refined skill body from stdin")?,
-            };
-            let (version, archived) =
-                skill::refine(&name, &body, description.as_deref(), when.as_deref())?;
-            println!(
-                "{} '{name}' → v{version} (prior version archived at {})",
-                style("refined").color256(splash::ACCENT),
-                archived.display()
-            );
-            Ok(())
-        }
-        SkillCmd::Fetch { url, name } => run_skill_fetch(&url, name.as_deref()).await,
-        SkillCmd::Search { query, limit } => {
-            let q = query.join(" ");
-            if q.trim().is_empty() {
-                anyhow::bail!(
-                    "a search query is required, e.g. `aizen skill search deploy fastapi`"
-                );
-            }
-            let hits = skill_registry::search(&q, limit.unwrap_or(20).clamp(1, 50)).await?;
-            if hits.is_empty() {
-                println!(
-                    "no skills on {} match '{q}'",
-                    skill_registry::registry_base()
-                );
-                return Ok(());
-            }
-            println!(
-                "{}",
-                style(format!(
-                    "{} result(s) from {} — install with `aizen skill install <owner/name>`",
-                    hits.len(),
-                    skill_registry::registry_base()
-                ))
-                .dim()
-            );
-            for sk in &hits {
-                println!("{}", sk.summary_line());
-            }
-            Ok(())
-        }
-        SkillCmd::Install { slug } => {
-            let sk = skill_registry::install(&slug).await?;
-            println!(
-                "{} '{}' → {}",
-                style("installed").color256(splash::ACCENT),
-                sk.name,
-                skill::skills_dir()
-                    .join(format!("{}.md", skill::sanitize_name(&sk.name)))
-                    .display()
-            );
-            Ok(())
-        }
-    }
-}
-
-/// GET a markdown skill from `url` and save it (name from `--name` > frontmatter > URL filename).
-/// SSRF-guarded like every other outbound fetch: the URL passes the net_guard floor and the fetch
-/// goes through the shared guarded client (no auto-redirects; every hop re-vetted; body bounded).
-pub(crate) async fn run_skill_fetch(url: &str, name_override: Option<&str>) -> Result<()> {
-    if !(url.starts_with("http://") || url.starts_with("https://")) {
-        anyhow::bail!("fetch needs an absolute http(s) URL");
-    }
-    crate::core::net_guard::guard_url_async(url).await?;
-    let http = crate::agent::reach::http::client()?;
-    let resp = crate::agent::reach::http::get(&http, url, &[])
-        .await
-        .with_context(|| format!("GET {url}"))?;
-    if !resp.is_success() {
-        anyhow::bail!("upstream returned HTTP {}", resp.status);
-    }
-    let text = resp.text();
-    // Fallback name from the URL's filename (strip a trailing .md).
-    let stem = url
-        .trim_end_matches('/')
-        .rsplit('/')
-        .next()
-        .unwrap_or("skill");
-    let stem = stem
-        .split(['?', '#'])
-        .next()
-        .unwrap_or(stem)
-        .trim_end_matches(".md");
-    let sk = skill::parse_markdown(&text, stem);
-    let name = name_override.unwrap_or(&sk.name);
-    let path = skill::save(name, &sk.description, &sk.when, &sk.body)?;
-    println!("fetched skill '{name}' → {}", path.display());
-    Ok(())
-}
-
-fn read_stdin(ctx: &'static str) -> Result<String> {
-    use std::io::Read;
-    let mut buf = String::new();
-    std::io::stdin().read_to_string(&mut buf).context(ctx)?;
-    // Strip a leading UTF-8 BOM (PowerShell's `|` prepends one) before trimming.
-    Ok(buf
-        .strip_prefix('\u{FEFF}')
-        .unwrap_or(&buf)
-        .trim()
-        .to_string())
-}
-
-// ── `aizen agents …` — the specialist sub-agent library (agency-agents format) ──
-
-/// A classified install source for `aizen agents install`.
-#[derive(Debug, PartialEq)]
-enum InstallSource {
-    /// `owner/repo` → cloned from github.com.
-    GitHubShorthand(String),
-    /// A full git URL (https `.git`/repo, `git@…`, `ssh://…`).
-    GitUrl(String),
-    /// A single `.md` agent file fetched over http(s).
-    FileUrl(String),
-    /// A local directory tree.
-    LocalDir(std::path::PathBuf),
-}
-
-/// Classify an install source string. Pure (no IO except an existing-dir probe) so it's unit-testable.
-fn classify_source(raw: &str) -> Result<InstallSource> {
-    let s = raw.trim();
-    if s.is_empty() {
-        anyhow::bail!("an install source is required (owner/repo, a git/.md URL, or a local dir)");
-    }
-    // http(s): a single .md file vs a git repo.
-    if s.starts_with("http://") || s.starts_with("https://") {
-        let path_only = s.split(['?', '#']).next().unwrap_or(s);
-        if path_only.to_ascii_lowercase().ends_with(".md") {
-            return Ok(InstallSource::FileUrl(s.to_string()));
-        }
-        return Ok(InstallSource::GitUrl(s.to_string()));
-    }
-    // scp-like / ssh git URLs, or any bare `*.git`.
-    if s.starts_with("git@") || s.starts_with("ssh://") || s.ends_with(".git") {
-        return Ok(InstallSource::GitUrl(s.to_string()));
-    }
-    // Explicit local-path forms, a Windows drive (`C:\…`), or an existing directory.
-    let drive = {
-        let b = s.as_bytes();
-        b.len() >= 2 && b[0].is_ascii_alphabetic() && b[1] == b':'
-    };
-    let looks_local = s.starts_with("./")
-        || s.starts_with("../")
-        || s.starts_with(".\\")
-        || s.starts_with("..\\")
-        || s.starts_with('/')
-        || s.starts_with('\\')
-        || drive;
-    if looks_local || std::path::Path::new(s).is_dir() {
-        return Ok(InstallSource::LocalDir(std::path::PathBuf::from(s)));
-    }
-    // GitHub shorthand: exactly `owner/repo` (one slash, both halves non-empty, no whitespace).
-    let parts: Vec<&str> = s.split('/').collect();
-    if parts.len() == 2
-        && parts
-            .iter()
-            .all(|p| !p.is_empty() && !p.contains(char::is_whitespace))
-    {
-        return Ok(InstallSource::GitHubShorthand(s.to_string()));
-    }
-    anyhow::bail!(
-        "unrecognized source '{raw}' — use owner/repo, an https git URL, a `.md` URL, or a local directory"
-    );
-}
-
-async fn run_agents(cmd: Option<AgentsCmd>) -> Result<()> {
-    match cmd {
-        None => {
-            agents_default_view();
-            Ok(())
-        }
-        Some(AgentsCmd::List {
-            division,
-            source,
-            enabled,
-            json,
-        }) => agents_list(division.as_deref(), source.as_deref(), enabled, json),
-        Some(AgentsCmd::Show { name }) => match agents::load(&name) {
-            Some(def) => {
-                println!("{}", agents::render_card(&def));
-                Ok(())
-            }
-            None => anyhow::bail!("no agent named '{name}' (try `aizen agents list`)"),
-        },
-        Some(AgentsCmd::Where) => {
-            agents_where();
-            Ok(())
-        }
-        Some(AgentsCmd::Install {
-            source,
-            yes,
-            enable_all,
-            as_name,
-        }) => agents_install(&source, yes, enable_all, as_name.as_deref()).await,
-        Some(AgentsCmd::Remove { name }) => {
-            if agents::delete_home(&name)? {
-                let _ = agents::set_enabled(&name, false);
-                println!("removed '{name}' from ~/.aizen/agents and unpinned it");
-            } else {
-                println!("(no agent named '{name}' under ~/.aizen/agents — `aizen agents where` shows the dirs)");
-            }
-            Ok(())
-        }
-        Some(AgentsCmd::Enable { name, all }) => agents_set_enabled(name.as_deref(), all, true),
-        Some(AgentsCmd::Disable { name, all }) => agents_set_enabled(name.as_deref(), all, false),
-        Some(AgentsCmd::SetProvider {
-            name,
-            provider,
-            model,
-            clear,
-        }) => {
-            let mut cfg = cli_config::load();
-            if agents::load(&name).is_none() {
-                anyhow::bail!("no agent named '{name}' (try `aizen agents list`)");
-            }
-            if clear {
-                cfg.set_agent_route(&name, None, None)?;
-                cli_config::save(&cfg)?;
-                println!(
-                    "{} cleared provider assignment on '{}'",
-                    crate::ui::theme::ok("✓"),
-                    name
-                );
-            } else {
-                let provider = provider
-                    .as_deref()
-                    .map(str::trim)
-                    .filter(|s| !s.is_empty())
-                    .context("pass a saved provider name (or --clear)")?;
-                cfg.set_agent_route(&name, Some(provider.to_string()), model)?;
-                cli_config::save(&cfg)?;
-                let route = cfg.agent_route(&name).expect("route just saved");
-                println!(
-                    "{} assigned '{}' to provider {} · model {}",
-                    crate::ui::theme::ok("✓"),
-                    name,
-                    route.provider.as_deref().unwrap_or("inherit"),
-                    route.model.as_deref().unwrap_or("provider default")
-                );
-            }
-            Ok(())
-        }
-        Some(AgentsCmd::SetModel { name, model, clear }) => {
-            let new_model = if clear {
-                None
-            } else {
-                match model.as_deref().map(str::trim).filter(|s| !s.is_empty()) {
-                    Some(m) => Some(m.to_string()),
-                    None => anyhow::bail!("pass a model id (or --clear to remove the pin)"),
-                }
-            };
-            let path = agents::set_model(&name, new_model.as_deref())?;
-            match &new_model {
-                Some(m) => println!(
-                    "{} pinned '{name}' to model {} ({})",
-                    crate::ui::theme::ok("✓"),
-                    style(m).color256(splash::ACCENT),
-                    path.display()
-                ),
-                None => println!(
-                    "{} cleared the model pin on '{name}' ({})",
-                    crate::ui::theme::ok("✓"),
-                    path.display()
-                ),
-            }
-            Ok(())
-        }
-    }
-}
-
-/// Bare `aizen agents`: list when any exist, else the install nudge.
-fn agents_default_view() {
-    if agents::has_any() {
-        let _ = agents_list(None, None, false, false);
-    } else {
-        agents_nudge();
-    }
-}
-
-fn agents_nudge() {
-    println!("No specialist agents yet. Install the agency-agents library with:");
-    println!(
-        "  {}",
-        style("aizen agents install msitarzewski/agency-agents").color256(splash::ACCENT)
-    );
-    println!("…or drop `.md` personas into ~/.aizen/agents (or ~/.claude/agents).");
-}
-
-fn agents_list(
-    division: Option<&str>,
-    source: Option<&str>,
-    enabled_only: bool,
-    json: bool,
-) -> Result<()> {
-    let enabled = agents::enabled_set();
-    let is_enabled = |slug: &str| enabled.as_ref().map(|e| e.contains(slug)).unwrap_or(false);
-
-    let mut all = agents::list();
-    if let Some(d) = division {
-        let d = d.to_lowercase();
-        all.retain(|a| a.division.as_deref() == Some(d.as_str()));
-    }
-    if let Some(src) = source {
-        let src = src.to_lowercase();
-        all.retain(|a| a.source.label() == src);
-    }
-    if enabled_only {
-        all.retain(|a| is_enabled(&a.slug()));
-    }
-
-    let cfg = cli_config::load();
-    if json {
-        let arr: Vec<serde_json::Value> = all
-            .iter()
-            .map(|a| {
-                serde_json::json!({
-                    "slug": a.slug(),
-                    "name": a.name,
-                    "description": a.description,
-                    "division": a.division,
-                    "source": a.source.label(),
-                    "model": a.model,
-                    "provider": cfg.agent_route(&a.slug()).and_then(|r| r.provider.clone()),
-                    "route_model": cfg.agent_route(&a.slug()).and_then(|r| r.model.clone()),
-                    "tools": a.tools,
-                    "enabled": is_enabled(&a.slug()),
-                    "path": a.source_path.display().to_string(),
-                })
-            })
-            .collect();
-        println!("{}", serde_json::to_string_pretty(&arr)?);
-        return Ok(());
-    }
-
-    if all.is_empty() {
-        if agents::has_any() {
-            println!("(no agents match the filter)");
-        } else {
-            agents_nudge();
-        }
-        return Ok(());
-    }
-
-    let mut by_div: std::collections::BTreeMap<String, Vec<&agents::AgentDef>> =
-        std::collections::BTreeMap::new();
-    for a in &all {
-        by_div
-            .entry(
-                a.division
-                    .clone()
-                    .unwrap_or_else(|| "(no division)".to_string()),
-            )
-            .or_default()
-            .push(a);
-    }
-    let total = all.len();
-    let enabled_count = all.iter().filter(|a| is_enabled(&a.slug())).count();
-    for (div, items) in &by_div {
-        println!("{}", style(format!("{div} ({})", items.len())).bold());
-        for a in items {
-            let mark = if is_enabled(&a.slug()) {
-                style("●").color256(splash::ACCENT).to_string()
-            } else {
-                style("○").dim().to_string()
-            };
-            let desc: String = a.description.chars().take(80).collect();
-            let route = cfg.agent_route(&a.slug());
-            let route_hint = route
-                .map(|r| {
-                    format!(
-                        " · provider {} · {}",
-                        r.provider.as_deref().unwrap_or("inherit"),
-                        r.model.as_deref().unwrap_or("default model")
-                    )
-                })
-                .unwrap_or_default();
-            println!(
-                "  {} {}  —  {}{}",
-                mark,
-                a.slug(),
-                desc.replace('\n', " "),
-                route_hint
-            );
-        }
-    }
-    let hint = if enabled.is_some() {
-        format!("{total} agent(s) · {enabled_count} pinned to <agents>. Dispatch: task(agent=\"<slug>\").")
-    } else {
-        format!("{total} agent(s) · none pinned — `aizen agents enable <slug>` to advertise them.")
-    };
-    println!("{}", style(hint).dim());
-    Ok(())
-}
-
-fn agents_where() {
-    println!("Specialist agent sources (lower → higher precedence):");
-    for (src, dir, n) in agents::source_counts() {
-        let status = if !dir.exists() {
-            style("(absent)").dim().to_string()
-        } else {
-            format!("{n} agent(s)")
-        };
-        println!("  {:<16} {}  [{}]", src.label(), dir.display(), status);
-    }
-    println!(
-        "{}",
-        style(
-            "Installs write to ~/.aizen/agents; a higher-precedence dir wins on a slug collision."
-        )
-        .dim()
-    );
-}
-
-fn agents_set_enabled(name: Option<&str>, all: bool, on: bool) -> Result<()> {
-    if all {
-        agents::set_all_enabled(on)?;
-        println!(
-            "{} all agents {} the <agents> index",
-            if on { "pinned" } else { "unpinned" },
-            if on { "to" } else { "from" }
-        );
-        return Ok(());
-    }
-    let name = name.context("provide an agent name, or pass --all")?;
-    let def = agents::load(name)
-        .with_context(|| format!("no agent named '{name}' (try `aizen agents list`)"))?;
-    agents::set_enabled(&def.slug(), on)?;
-    println!(
-        "{} '{}' {} the <agents> index",
-        if on { "pinned" } else { "unpinned" },
-        def.slug(),
-        if on { "to" } else { "from" }
-    );
-    Ok(())
-}
-
-fn confirm_write(prompt: &str) -> Result<bool> {
-    Ok(Confirm::with_theme(&ui_theme())
-        .with_prompt(prompt)
-        .default(true)
-        .interact_opt()
-        .ok()
-        .flatten()
-        .unwrap_or(false))
-}
-
-/// A filesystem-safe directory name from a repo slug/URL (last segment, `.git` stripped).
-fn sanitize_repo_name(s: &str) -> String {
-    let base = s
-        .trim_end_matches('/')
-        .rsplit(['/', ':', '\\'])
-        .next()
-        .unwrap_or(s);
-    let base = base.trim_end_matches(".git");
-    let cleaned: String = base
-        .chars()
-        .map(|c| {
-            if c.is_alphanumeric() || c == '-' || c == '_' || c == '.' {
-                c
-            } else {
-                '-'
-            }
-        })
-        .collect();
-    let cleaned = cleaned.trim_matches(['-', '.']).to_string();
-    if cleaned.is_empty() {
-        "agents".to_string()
-    } else {
-        cleaned
-    }
-}
-
-fn unique_n() -> u64 {
-    use std::sync::atomic::{AtomicU64, Ordering};
-    static N: AtomicU64 = AtomicU64::new(0);
-    N.fetch_add(1, Ordering::Relaxed)
-}
-
-/// `git clone --depth 1` (shallow, quiet, NO submodule recursion) into `dest`. Cloning runs NO repo
-/// code — it only reads files. `--no-recurse-submodules` stops a hostile `.gitmodules` from making git
-/// fetch arbitrary (possibly internal) submodule URLs we never vetted.
-fn git_clone_shallow(url: &str, dest: &std::path::Path) -> Result<()> {
-    let out = crate::core::gitx::command()?
-        .args([
-            "clone",
-            "--depth",
-            "1",
-            "--no-recurse-submodules",
-            "--quiet",
-            url,
-        ])
-        .arg(dest)
-        .output()
-        .context("running `git clone` (is git installed and on PATH?)")?;
-    if !out.status.success() {
-        anyhow::bail!(
-            "git clone failed: {}",
-            String::from_utf8_lossy(&out.stderr).trim()
-        );
-    }
-    Ok(())
-}
-
-/// Extract the host from a NON-http(s) git URL (`git@host:path`, `ssh://[user@]host[:port]/path`,
-/// `git://host/path`) so the SSRF floor can guard it too. `None` if no host is discernible.
-fn git_url_host(url: &str) -> Option<String> {
-    let non_empty = |s: &str| {
-        let s = s.trim();
-        if s.is_empty() {
-            None
-        } else {
-            Some(s.to_string())
-        }
-    };
-    // scp-like: [user@]host:path (no scheme).
-    if url.starts_with("git@") || (url.contains('@') && url.contains(':') && !url.contains("://")) {
-        let after_at = url.rsplit('@').next().unwrap_or(url);
-        return non_empty(after_at.split(':').next().unwrap_or(after_at));
-    }
-    let rest = url
-        .strip_prefix("ssh://")
-        .or_else(|| url.strip_prefix("git://"))?;
-    let after_at = rest.rsplit('@').next().unwrap_or(rest);
-    non_empty(after_at.split(['/', ':']).next().unwrap_or(after_at))
-}
-
-/// Copy every `*.md` that `looks_like_agent` from `src` (recursively, dotdirs skipped) into
-/// `dest_root`, preserving the relative subpath. Returns `(copied, skipped)`.
-fn copy_agent_tree(src: &std::path::Path, dest_root: &std::path::Path) -> Result<(usize, usize)> {
-    let mut copied = 0;
-    let mut skipped = 0;
-    copy_agent_walk(src, src, dest_root, &mut copied, &mut skipped, 0)?;
-    Ok((copied, skipped))
-}
-
-fn copy_agent_walk(
-    dir: &std::path::Path,
-    src_root: &std::path::Path,
-    dest_root: &std::path::Path,
-    copied: &mut usize,
-    skipped: &mut usize,
-    depth: usize,
-) -> Result<()> {
-    if depth > 12 {
-        return Ok(());
-    }
-    let Ok(rd) = std::fs::read_dir(dir) else {
-        return Ok(());
-    };
-    for e in rd.flatten() {
-        let p = e.path();
-        let name = e.file_name().to_string_lossy().to_string();
-        if name.starts_with('.') {
-            continue; // skip .git, dotfiles
-        }
-        // Never follow symlinks out of an UNTRUSTED cloned tree (a symlinked `x.md` could pull in a
-        // file outside the repo; a symlinked dir could escape it). The loader treats the user's own
-        // dirs as trusted, but install copies third-party content, so refuse symlinks here.
-        if e.file_type().map(|t| t.is_symlink()).unwrap_or(false) {
-            continue;
-        }
-        if e.file_type().map(|t| t.is_dir()).unwrap_or(false) {
-            copy_agent_walk(&p, src_root, dest_root, copied, skipped, depth + 1)?;
-        } else if p
-            .extension()
-            .and_then(|x| x.to_str())
-            .is_some_and(|x| x.eq_ignore_ascii_case("md"))
-        {
-            let Ok(content) = std::fs::read_to_string(&p) else {
-                continue;
-            };
-            if !agents::looks_like_agent(&content) {
-                *skipped += 1;
-                continue;
-            }
-            let rel = p.strip_prefix(src_root).unwrap_or(&p);
-            let dest = dest_root.join(rel);
-            if let Some(parent) = dest.parent() {
-                std::fs::create_dir_all(parent)
-                    .with_context(|| format!("creating {}", parent.display()))?;
-            }
-            std::fs::write(&dest, content)
-                .with_context(|| format!("writing {}", dest.display()))?;
-            *copied += 1;
-        }
-    }
-    Ok(())
-}
-
-async fn agents_install(
-    source: &str,
-    yes: bool,
-    enable_all: bool,
-    as_name: Option<&str>,
-) -> Result<()> {
-    let classified = classify_source(source)?;
-    println!(
-        "{}",
-        style("⚠ agent bodies are third-party system prompts — they run as sub-agents with edit/shell scope. Review before pinning.").dim()
-    );
-    match classified {
-        InstallSource::FileUrl(url) => {
-            crate::core::net_guard::guard_url_async(&url).await?;
-            if !yes && !confirm_write(&format!("Fetch and install the agent at {url}?"))? {
-                println!("cancelled.");
-                return Ok(());
-            }
-            // Fetch through the shared guarded client (auto-redirects OFF, every hop re-vetted
-            // against the net_guard floor) — a plain reqwest client follows up to 10 redirects and
-            // would re-vet only the first hop, so a 302 → 169.254.169.254 / localhost slips through.
-            let http = crate::agent::reach::http::client()?;
-            let resp = crate::agent::reach::http::get(&http, &url, &[])
-                .await
-                .with_context(|| format!("GET {url}"))?;
-            if !resp.is_success() {
-                anyhow::bail!("upstream returned HTTP {}", resp.status);
-            }
-            let text = resp.text();
-            if !agents::looks_like_agent(&text) {
-                anyhow::bail!(
-                    "that URL isn't an agent (needs frontmatter `name:` + a non-empty body)"
-                );
-            }
-            let stem = url
-                .trim_end_matches('/')
-                .rsplit('/')
-                .next()
-                .unwrap_or("agent");
-            let stem = stem
-                .split(['?', '#'])
-                .next()
-                .unwrap_or(stem)
-                .trim_end_matches(".md");
-            let stem = if stem.is_empty() { "agent" } else { stem }; // e.g. a URL ending in "/.md"
-            let path = agents::save_home(&text, as_name.unwrap_or(stem))?;
-            println!("installed 1 agent → {}", path.display());
-            if enable_all {
-                agents::set_all_enabled(true)?;
-                println!("…and pinned all agents to <agents>.");
-            }
-            Ok(())
-        }
-        InstallSource::LocalDir(dir) => {
-            if !dir.is_dir() {
-                anyhow::bail!("not a directory: {}", dir.display());
-            }
-            if !yes && !confirm_write(&format!("Install agents from {}?", dir.display()))? {
-                println!("cancelled.");
-                return Ok(());
-            }
-            let label = dir.file_name().and_then(|s| s.to_str()).unwrap_or("local");
-            let dest = agents::agents_dir().join(sanitize_repo_name(label));
-            std::fs::create_dir_all(&dest)
-                .with_context(|| format!("creating {}", dest.display()))?;
-            let (copied, skipped) = copy_agent_tree(&dir, &dest)?;
-            crate::core::config::harden_dir(&agents::agents_dir());
-            finish_install(copied, skipped, &dest, enable_all)
-        }
-        InstallSource::GitHubShorthand(slug) => {
-            install_from_git(
-                &format!("https://github.com/{slug}.git"),
-                &slug,
-                yes,
-                enable_all,
-            )
-            .await
-        }
-        InstallSource::GitUrl(url) => {
-            let label = sanitize_repo_name(&url);
-            install_from_git(&url, &label, yes, enable_all).await
-        }
-    }
-}
-
-async fn install_from_git(url: &str, label: &str, yes: bool, enable_all: bool) -> Result<()> {
-    // SSRF floor — guard the destination host whatever the git transport is.
-    if url.starts_with("https://") || url.starts_with("http://") {
-        crate::core::net_guard::guard_url_async(url).await?;
-    } else if let Some(host) = git_url_host(url) {
-        // ssh:// / git@ / git:// never go through the http(s) guard; guard the resolved host directly
-        // so an internal endpoint (e.g. git@10.0.0.5:…) can't be reached past the floor.
-        crate::core::net_guard::guard_url_async(&format!("https://{host}")).await?;
-    }
-    if !yes
-        && !confirm_write(&format!(
-            "Clone {url} and install its agents into ~/.aizen/agents?"
-        ))?
-    {
-        println!("cancelled.");
-        return Ok(());
-    }
-    let repo = sanitize_repo_name(label);
-    let dest = agents::agents_dir().join(&repo);
-    let staging = std::env::temp_dir().join(format!(
-        "aizen-agents-clone-{}-{}",
-        std::process::id(),
-        unique_n()
-    ));
-    let _ = std::fs::remove_dir_all(&staging);
-    println!("{}", style(format!("cloning {url} …")).dim());
-
-    let url_s = url.to_string();
-    let staging_c = staging.clone();
-    let clone_res = tokio::task::spawn_blocking(move || git_clone_shallow(&url_s, &staging_c))
-        .await
-        .context("clone task panicked")?;
-
-    // Always clean the staging clone, whether or not the copy succeeds.
-    let outcome = (|| -> Result<(usize, usize)> {
-        clone_res?;
-        std::fs::create_dir_all(&dest).with_context(|| format!("creating {}", dest.display()))?;
-        let counts = copy_agent_tree(&staging, &dest)?;
-        crate::core::config::harden_dir(&agents::agents_dir());
-        Ok(counts)
-    })();
-    let _ = std::fs::remove_dir_all(&staging);
-
-    let (copied, skipped) = outcome?;
-    finish_install(copied, skipped, &dest, enable_all)
-}
-
-fn finish_install(
-    copied: usize,
-    skipped: usize,
-    dest: &std::path::Path,
-    enable_all: bool,
-) -> Result<()> {
-    if copied == 0 {
-        // Nothing landed. Use `remove_dir` (empty-only), NOT `remove_dir_all`: it removes the dir we
-        // just created but can never wipe pre-existing user files in a same-named directory.
-        let _ = std::fs::remove_dir(dest);
-        anyhow::bail!(
-            "no agents found ({skipped} non-agent file(s) skipped) — the source had no `*.md` with frontmatter `name:` + a body"
-        );
-    }
-    println!(
-        "installed {copied} agent(s) → {} ({skipped} non-agent file(s) skipped)",
-        dest.display()
-    );
-    if enable_all {
-        agents::set_all_enabled(true)?;
-        println!("…and pinned all agents to <agents>.");
-    } else {
-        println!(
-            "{}",
-            style("none are pinned yet — `aizen agents enable <slug>` (or re-run with --enable-all) to advertise them.").dim()
-        );
-    }
-    println!(
-        "{}",
-        style("review: `aizen agents list` · `aizen agents show <slug>`").dim()
-    );
     Ok(())
 }
