+++ src/tests/main_suite.rs
@@ -2,6 +2,7 @@
 //! Kept as a `#[path]` child module so `use super::*` still resolves to the crate root.
 
 use super::*;
+use anyhow::anyhow; // `main.rs` no longer needs it now that `apps` moved out; these tests still do.
 
 #[test]
 fn config_provider_subcommands_parse() {
