+++ src/ui/mod.rs
@@ -6,6 +6,7 @@ pub mod cards;
 pub mod channel_markdown;
 pub mod config_ui;
 pub mod context_report;
+pub mod effort_ui;
 pub mod icons;
 pub mod image_input;
 pub mod links;
@@ -13,6 +14,7 @@ pub mod markdown;
 pub mod menus;
 pub mod mermaid;
 pub mod moonscape;
+pub mod plain_input;
 pub mod spinner;
 pub mod splash;
 pub mod theme;
